[roll] Roll infra/recipes [nullcontext] Delete
`nullcontext` is part of the standard library as of Python 3.7, so we
can use the standard library function instead of our own custom recipe
module.

Also switch most context calls to call the context before the `with`
statement to avoid unnecessary lambdas. The setup logic for a context
manager is generally implemented in the `__enter__` method of the object
returned by `__call__`, and `__enter__` is only called when the `with`
statement executes, so this is a no-op.

Original-Bug: 83858
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/590642
Original-Revision: 602705546e4c58fcc587475d8dcf1eb5d03d7bf1
GitOrigin-RevId: 252b067a95aba2cb18b8b53632d6cc940fa8df39
Change-Id: I7e8e59e0275e96609680190917dbb71dd8f14ee5