[roll] Roll fuchsia [hypervisor] Always cancel interrupts on kick
When kicking a VCPU, there is a possibility that it may be in the
process of executing a WFI or HLT instruction and have left the guest
but not currently be blocked on a WaitQueue. Because of this, we should
always attempt to cancel any pending wait-for-interrupts.

Manually verified this fix by running hypervisor-tests-pkg 1000 times on
a Sherlock with no failure of Guest.VcpuSuspendThread, which is
particularly good at picking up this problem.

Multiply: hypervisor-tests-pkg
Original-Fixed: 86017
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590328
Original-Revision: 5777a058ddceefdd8514ec864278fc12e2b7ae00
GitOrigin-RevId: d7c102c2c1ec4da79903efc3df623c7e4ee520b1
Change-Id: I830f54e17c1106b373cc5fc0935712f341ae92b1