[roll] Roll fuchsia [fshost] Keep the reboot reason cache file
The Copier only allows for excluding files when copying and it's a lot
easier to exclude all of the cache than to exclude everything except for
the reboot reason file. The reboot reason file is read separately and
inserted into the copier so it is included in the data size estimate and
written with the rest of the files.

Original-Bug: 84885
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589929
Original-Revision: f7118f38778f890f0ccd7809fe8ea9f4dc9c7f96
GitOrigin-RevId: 087f6ac63af14afb945ba5ae7a0a4289fe9229e0
Change-Id: I201fda7663b81bdd026c6609c995d441af236832