[roll] Roll fuchsia [kernel][uart] Make RX mask operations atomic with console buffer checks
The pl011 driver masks RX interrupts when the console buffer is full,
thus disallowing any further characters to be read in from the RX line.
The interrupts are re-enabled when getc is called. This can lead to the
following race condition:
< let's assume the console buffer is full to start >
1. Thread A receives an RX interrupt and enters the interrupt handler.
   The full check returns true, but before we mask interrupts...
2. Thread B runs getc, reads a character, and unmasks interrupts.
3. Thread A resumes and masks interrupts.
This leaves us in a state where the console buffer has space but is not
receiving RX interrupts. This CL solves the issue by acquiring the UART
spinlock before performing the console buffer full check and masking
interrupts, as well as before unmasking interrupts. This should solve
the attached bug, which I've validated by running >1000 runs of
bringup.arm64 with 0 flakes. The previous flake rate was 1/400.

Original-Bug: 80933
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590482
Original-Revision: ee212548ca3ee684ae1a34d288ea1bf8a5260a90
GitOrigin-RevId: ab98928a7c03cc888573ed403ae2622f7b2601ef
Change-Id: I724577ddcc58b3a6e722dad3298e19af431e25d5