[roll] Roll fuchsia [wlan][brcmfmac] Override incorrect reason code
On certain occations, we notice the firmware returning a failure
status for an assoc request, but with a success reason code. This
results in the WLAN stack incorrectly thinking that the connection
was successful.

The below mentioned bug is currently tracking this issue. In the
meantime, this will serve as a work around by overriding the reason
code with a generic error under these circumstances.

Original-Bug: 85446
Test: fx test assoc_test_component
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588833
Original-Revision: fc0c6f9513c1e8ac938e1fc6cb0bc4ee5bc00c48
GitOrigin-RevId: 14457661e44a039a2cb916f28a59cfd45c4efe7e
Change-Id: Iba5dd26f3917a41d6117d6bebeccdaf07508e6f4