[roll] Roll fuchsia [kernel][vm] Correct PagerCounts
Previously it was assumed there would always be the full number of
buckets in use and so the oldest queue would be kNumPagerBacked from
the newest queue. Now that the LRU queue might get processed in
preparation for aging to have a queue ready for it this can result
in the oldest count being falsely reported as zero. This fixes by
always considering the LRU index to be the oldest.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591061
Original-Revision: 72c341c9760ae0f15650b8ce6e8d52ab2968beb4
GitOrigin-RevId: 87a21146f50722e6a02660f3b0e6c93810e0706e
Change-Id: I42cb1991bab31418c527a39b586b4fc0acc436e7