[roll] Roll fuchsia [clkctl][aml-clk] Implement enable/disable via clkctl
For diagnostic and debugging purposes it's sometimes useful to be able
to enable and disable clocks from the command line.

This patch adds an enable/disable interface to the clock-impl interface
and adds functionality to clkctl to manipulate that interface.

Original-Bug: 85139
Test: Started up a peripheral and disabled its clock. Observed that
the peripheral stopped sending data. Enabled the clock again and
observed that the peripheral started working again.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/586709
Original-Revision: 7ded68be2e5d92cef7b547cec8235b66c60e2908
GitOrigin-RevId: b4385dcfef16fc0b732864e90c42ef62e5bda4a1
Change-Id: I19ffdf4b07e8b136755cccbab425d4ac2b367678