[roll] Roll fuchsia [fx] Suggest sync-to before repro'ing old build
Previously, if an infra build used a fint parameters file that no longer
exists at HEAD (often because the file is from an old build or old
release branch and uses an old naming scheme), `fx repro` would fail
with an uninformative error message about the file not existing.

Now `fx repro` will still fail, but it will suggest first running `fx
sync-to` to go back to the revision used by the build, where the fint
params file should exist.

I also cleaned up some backwards compatibility shims that were supposed
to make it possible to reproduce builds that used fint params with
legacy naming schemes. The shims depended on the user running `fx
sync-to` before trying to run the repro `fx set` and `fx build` commands
– otherwise the file wouldn't be available. So it's simpler to
*immediately* fail with an informative error message if the user must to
run `fx sync-to` to go back to the version of the checkout used by the
build. The version of `fx repro` in that old version of the checkout
should be compatible with the build, so we don't need to maintain
support for old naming schemes at HEAD.

Original-Bug: 85773
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588303
Original-Revision: 31fcfe0145108076ce412937194987f655a53c8a
GitOrigin-RevId: a98e25f66d054b5ae09c1dd4246b3f2db040316d
Change-Id: I493731844f5d0bf07a712870bfccb554720e6a1c