[roll] Roll fuchsia [kernel][test] Fix kstack-mp-sync-exec test
This change fixes a problem with kstack-mp-sync-exec test.  Prior to
this change, the two threads of the test could end up running on the
same CPU, which could cause the test to take tens of seconds to
complete because we'd have to wait and get lucky so that the
non-pinned thread would be moved to some other CPU.

Update the test to use two separate CPUs with each thread pinned to a
different one.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590173
Original-Revision: 91c1c7b350f13a3b46c00d0eff5295137789a2d5
GitOrigin-RevId: d59ebf30f2b4a2c67857942c517249db83090309
Change-Id: I11476ff51a4ad65e3697be574138ee720ef0e55e