[roll] Roll fuchsia [virtualization][vsock] Directly send reset packets to guests
Previously, when VirtioVsock would receive an invalid packet from the
guest it would create a new NullConnection to track the need to send
a reset back. This would then be processed next time the transmit path
(Mux) was called.

In this CL, we implement a send queue, and use this to directly send the
reset if descriptors are available. If not, instead of tracking the
reset in the form of a fake connection, we use a packet queue.

The send queue will be particularly useful in upcoming CLs where we want
to send control packets back to the guest from within a connection
without having to modify its state.

Multiply: virtualization-core-tests, vmm_unittests
Original-Bug: fxbug.dev/85702

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584725
Original-Revision: 586c9730780d0715172732856231b922d0322b8e
GitOrigin-RevId: 2da494bcb6bb94ec3d6cc851c3539fc3f330ca3d
Change-Id: I39606a9d37d249622fc12e0c732dd498457a2aba