[roll] Roll fuchsia [session_manager] Don't panic on error while proxying.
It doesn't make sense for an error in a proxied service to take down
the whole session_manager.  It's up to the client of the service to
decide how to handle failures, just as if they were connecting to the
service directly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591056
Original-Revision: f52e8b9e4728a773be2677e351096ff5e6d7d3c5
GitOrigin-RevId: 3fcec6ef42fb9943d41e9faa5fefe4096156dfae
Change-Id: Ic5e8c0e9269dbcbaed2f485c8a5949a9f8101d4b