[roll] Roll fuchsia [fuchsia.sys2] Change facets to Dictionary.
The type for the `facets` in the `ComponentDecl`
table was changed to `fuchsia.data.Dictionary`.
This was done in order to bring this type to parity
with its SDK counterpart, which uses the `Dictionary`
instead of a custom type.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591365
Original-Revision: ffea9dd1ee828a0a8ee9dd61a1dd3112a1f0802f
GitOrigin-RevId: 3a967978a4487bf753af5e2be5b9281e90896bbb
Change-Id: I4e2d417063414bca8407c3c17b71295abe1c72d8