[roll] Roll fuchsia [netstack] Closing device closes DeviceControl
The logic of closing the DeviceControl channel when the device is closed
was not implemented correctly. The test
`device_control_closes_on_device_close` which was checking for it was
written incorrectly, and was observing a Netstack crash through the
channel closure instead of the expected DeviceControl channel closure.

Fix the logic and tighten the test.

Multiply: netstack-interfaces-admin-fidl-test: 30
Original-Bug: 85877
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591842
Original-Revision: 47967bacf34547c331d394595206de280ba9ae6e
GitOrigin-RevId: 45463f93279243dee9b5be4f6b953519e60fe248
Change-Id: I99eece817e27d94554587ce57f734b26afc7d88b