[roll] Roll fuchsia [ffx] Create Writer type
This creates a library at ffx/lib/writer which holds a Writer struct
that will be injected into plugins to handle output. The initial use
case is to support machine output but the end goal is that all writing
done by a plugin will run through this type.

This is the first in a series of CLs that fleshes out this concept and
introduces it into the codebase.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576633
Original-Revision: 8bf4d449de7965455446f179d495e19e65091b1a
GitOrigin-RevId: 11605656f6f2aa0a4ff436eacceb056e1920c882
Change-Id: Ibf4abf7857847112623e82fad05a26b0817491ab