[roll] Roll fuchsia [rust][clippy] Add flags to clippy script for tricium
The recipe that runs for tricium will build the clippy targets
for the set of changed files on its own. It will use this script
to perform that file to target mapping, and then when the analyzer
runs this script to fetch lints it will skip the ninja step because
the targets are already built.

Original-Bug: 86101
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590634
Original-Revision: 18060ac5f79f14d748bf2b37b82bf656a2df24b5
GitOrigin-RevId: 1dadd299cbdad6d2a647715eea83169479f1dbae
Change-Id: I04d82a09c5bb619d6cce1d6cebb9ad2f8b76e953