[roll] Roll fuchsia [lockdep] Fix edge cases for lockdep when used in the kernel.
When used in a kernel environment, lockdep has some edge cases which
need to be handled in a slightly special fashion.  Kernel code is
slightly special in that it can not only be interrupted at any point
in time due to an exception/interrupt, but the exception/interrupt
code can share locks with the kernel code which was executing
beforehand, which can lead to issues.

See the following for details:
http://fxb/84890
http://fxb/85289

To address these issues, the following changes have been made:

First, a PreValidate step has been added to the LockPolicy template
interface. This allows users to change the behavior for one of their
lock types to take actions immediately before validation happens as a
lock is acquired. This gives IrqSave spinlocks the ability to disable
interrupts before validation takes place, avoiding a potential race
which can lead to the detection of a false positive lockdep violation.

Second, a lock type's flags are now passed to the user supplied
SystemGetThreadLockState function.  This allows the kernel to choose a
different thread lock state based on the type of lock, avoiding the
context mix-up which lead to http://fxb/85289.

Test: new unit tests
Original-Fixed: 84890, 85289
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591514
Original-Revision: af291137e16880828301276e01b91ffd52f3e04b
GitOrigin-RevId: 73bfc24412da174dbd2068aaeb50189e61a686cd
Change-Id: I0c7037629df4357566ab25ad9f02b363485d3920