[roll] Roll fuchsia [ffx] Create machine flag and hook up with Writer
We add a `--machine <format>` flag to the top level ffx command with
json as the currently only supported format. This also adds the Writer
type into the set of things that can be injected into a plugin and hooks
up this flag to the construction of the Writer object that is owned by
the Injector.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577347
Original-Revision: 0e336b3c2116b36db0576470dd40af2f67f23b3a
GitOrigin-RevId: 832999294a48d4c8519e5714ca497a1b1d19d752
Change-Id: I6621c62e4cf4a1f9398b24d1ded83d5c36461e07