[roll] Roll fuchsia [ffx] Use machine flag in target list
This is a test case of the machine flag where we use it in the target
list command. I chose this because target list already has a format flag
with a json option so not much was needed within the plugin besides
grabbing the newly provided object and calling the relevant methods.
This still looks a bit rough because the Writer object is not managing
all output at this point. This gets cleaned up in a later CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576636
Original-Revision: 126c9ebc2334f9053cd64b7c669335bbc49c424a
GitOrigin-RevId: afc38ee7fd1bbcf52b328c03a64c261ff0b3554e
Change-Id: Ifbd8f61b7bb2cf85ca37821a367f2b9966eaab31