[roll] Roll fuchsia [ffx] Add line, info, and error methods to Writer
This adds methods to the Writer object which can be used to bridge the
gap between the current state of the world and the desired future where
most output is handled by producing one of a set of widgets that we
provide. The main method is `line` which is basically just println but
is tied together with the machine flag as well as a test mechanism.
Replacing calls to println with writer.line will allow for easy testing
and easy turning off of output when in machine mode.

The info and error methods write to stderr. I think we only need one of
these but I am still torn about what it call it. I think it should be
info as that is what people use it for. Real errors should be returned
as an error from the plugin so I don't think errors should be reported
inline. But information can be reported to this other stream.

This also does the clean up for target list to stop using stdout/stderr
directly and instead running everything through the writer object. This
is indicative of what it would look like in other plugins. Functionally
nothing should change from the current state.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576637
Original-Revision: 1afb94e415ced43b4f581dff474c05b1aac13e26
GitOrigin-RevId: f2609f00d8e648df0b77a92a58ff4ff967cd310d
Change-Id: I5cebd4056313bab94a027f5ec6729112715f12fc