[roll] Roll fuchsia [brcmfmac] Fix SDIO control frame threading issue
When transmitting SDIO control frames, usually things like setting
iovars, there is a possibility that the thread waiting for the
transmission to complete will detect an error even if the transmission
went well. This happens because the atomic that indicates if the
transmission is complete or not is not set until AFTER the waiting
thread is woken up. Most of the time the thread scheduling allows the
transmitting thread to keep running until it sets the atomic but
sometimes we get unlucky and the waiting thread immediately gets to run.

Fix this by setting the variable BEFORE the waiting thread is woken up.

Original-Bug: 86222
Test: Manual testing
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591507
Original-Revision: e6cc189e28fed704449c5334180c1be0ea3912ae
GitOrigin-RevId: 75c1ff199c8481241200d7fd778a57e50a180acf
Change-Id: Ief3d3483490c73ffa05525ae37d70ff1316113e6