[roll] Roll fuchsia [brcmfmac] Make Device::Init return a status code
Currently the brcmfmac Device class has a pure virtual method that
subclasses implement to perform initialization. This method has an
InitTxn object as the parameter and the initialization is supposed to
call Reply on this object with the resulting status code. This prevents
the part of Device that calls Init from inspecting the result of the
subclasses Init. In order to make decisions about further initialization
the Device class needs to know what the result of Init was.

This changes the Init method to return a zx_status_t instead and the
Device class will instead use that status to call Reply on the InitTxn
object. No implementation of init did anything else with the InitTxn
object at this point.

Test: fx test brcmfmac_sim_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591289
Original-Revision: a63ceecaf1240847ebfd70af89a3ce4ad411bba9
GitOrigin-RevId: 701116237953a2af1d307c49cce3a01370b0b0de
Change-Id: Ia6063f27464327a995ed8ff91107a9ec28f33bd2