[roll] Roll fuchsia [ffx] Add schema plugin
This adds a plugin called `schema` which uses the generated code from
the attribute attached to the ffx_plugin macro to dump the schema for
all plugins.

This is a bit of a hack because of the organization of the code. The
top entry point for the generated code that is able to walk through all
the plugins has to live in ffx_lib_suite. However you can't depend on
that lib from inside a plugin because that lib already depends on all
the plugins because that is how we do dispatch. So I was left with a
choice on how to add a subcommand that actually kinda lives in the main
lib. I chose to create a normal plugin but to intercept that in main.rs
and not call into the implementation but to rather call the relevant
generated function. The other approach would be to hack the subcommand
generator to add a custom extra bit that does not come from the list of
plugins. Neither way seemed to be clearly better.

Nevertheless, the point of this subcommand is to be able to generate a
golden file from it and then compare that with a fresh run within a
test to stop changes that would break machine stable contracts.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582050
Original-Revision: 1387014ed111dc923a1b67bfa24eaf468db0e753
GitOrigin-RevId: 9f81e62ceaefe9af0dec3067bce2380b27574644
Change-Id: I79de2279d8ac4981e129084fb3cf6a605ab3a83e