[roll] Roll fuchsia [fx sync-to] Don't fail if no repos are dirty
We currently assume that if `jiri runp` emits any output that means some
repos are dirty, but this is not correct, as `jiri runp` emits a warning
to stdout if the user has not yet chosen to enable or disable
analystics. So we should only fail this check if the output actually
references any dirty repos.

Original-Bug: 68492
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591987
Original-Revision: 17b7af728e1ea061da63e619ddc650f132553110
GitOrigin-RevId: e965c5e01f41aee774443da007a8e67a31ac36c2
Change-Id: I64e8793844cb4fdde24648ff9dbc5e11833f7ec9