[roll] Roll fuchsia [fidlgen][go] go:embed templates, also cleanup
This moves fidlgen_go to using go:embed rather than strings for
embedding templates. Because the go formatter is part of the standard
library rather than an external utility that's being called the common
interface for formatters needed to be simplified. That resulted in a few
follow-on changes that ended up simplifying the writing of files in
general.

Generally the code moves away from constructing a series of pipes to
just reading and writing whole generated files. In practice this
basically what was happening, but it was obscured by complexity in the
interfaces.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589731
Original-Revision: de60385f02bc4b7fc9c1d0735094a7b13567639b
GitOrigin-RevId: 3197f55cd30e6621c65fea564c0cfc39d360db5d
Change-Id: I66e858507c49aff5c017c07b1bd883c9f316b5c8