[roll] Roll fuchsia [test_runners][elf] Refactor elf_test_runner.
This CL adds a ComponentLauncher trait to the elf_test_runner and makes
most of the runner's functionality available as rustc_library target.

The purpose here is to facilitate making variants of the elf_test_runner
that differ only in how the launch the component process. As an example,
the fuzz_test_runner described by RFC-0117 needs to install a startup
handle to a channel in the launched process, but is otherwise identical
to elf_test_runner. With this refactor, the fuzz_test_runner can simply
provide its own impl of ComponentLauncher.

Original-Bug: 24740
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/577305
Original-Revision: b8f255f7656e6a6557ea4a9bee03e4d0b7a2a755
GitOrigin-RevId: edef5a461a4a04ee7e1ebb71c6417e7de653bf20
Change-Id: I01a683d127a28ad155ae419a2696728506fd1b5e