[roll] Roll fuchsia [rust] add allows for holding over await point
Some types in the new version of rustc are marked as being
unsafe to hold over an await point.  These types are marked with
a must_not_suspend lint and produce a warning.

Allow the current usages in the tree explicitly, with a pointer to
the bug for a further explanation and next steps.

Original-Bug: 86245
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591287
Original-Revision: dfc41ac1bc16ff576864d19c4c00f7b01e75587d
GitOrigin-RevId: 60ea001ba9fc8898403a2c8ce8616bb9f2e14c28
Change-Id: I8170d66d7edb22c2d6614de2405c95cfa12b9823