[roll] Roll fuchsia [kernel][uart][pl011] Hold spinlock when masking/unmasking RX/TX
For the pl011 masking/unmasking RX/TX interrupts is not atomic because
it requires a register read, an OR or AND, and register write.
Because it's not naturally atomic it's possible for one thread to
stomp another's changes.

Change the pl011 uart driver to guard mask/unmask operations using the
uart spinlock.

Original-Bug: 80933
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591517
Original-Revision: 6d3256ba0e6be6be513a95b677e7180a9a44a61d
GitOrigin-RevId: c0e71d53cd0e9f6f2e125bd68908af36d2187434
Change-Id: Ifd20fcb84365d5bd595f4b7ee80b88ba7e2d3536