[roll] Roll fuchsia [vm-stress-test] Re-supply after err due to resize.
Overall this allows vm-stress-test to be green more often and
purple less often.  We were seeing vm-stress-test sometimes take
a long time to exit (minutes), with the kernel output complaining
about a user pager taking too long.

If the VMO is resized to be smaller between when the pager request
is generated and when it's handled by the user pager, and the
supply_pages ends up trying to supply a page beyond the shorter end
of the VMO, we can get ZX_ERR_OUT_OF_RANGE with zero pages supplied.

However, the kernel still expects the user pager to supply the pages
up to the end of the VMO, so get the vmo size before supplying pages
and keep trying until it works or fails with an unexpected error.

Tested: vm-stress-test with VMO_VALIDATION enabled
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591361
Original-Revision: 40908972f869386c1400a440402b1f5bcbd0b268
GitOrigin-RevId: 98370352e803bf01f32e2c45850f9511d10ba2ff
Change-Id: I791bc7e246f4ee7ac770a4eae036c6f9046f2225