[roll] Roll fuchsia [kernel][vm] Check for overflow in eviction target
Check for overflows when adding a new min pages target to the existing
one. In practice this should never happen as it implies we've been asked
to evict an unreasonably large amount of pages. This guards against
garbage values and bugs in target calculation arithmetic.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592022
Original-Revision: 6834fda54b9461f411887100b9f6cb174336fe98
GitOrigin-RevId: 4b6d12c453692b8643b64a3db2a1d67a843a687c
Change-Id: I7020396c634464585e5364462d6e129ab2bab893