[roll] Roll third_party/scudo/src [scudo] Skip AllocAfterFork test on machines with low max_map_count
Reducing the number of iterations in that test with D111342 helped,
but the failure still occured flakily when the test is ran as part
of a large test suite.

Reducing further the number of iterations might not be good enough,
so we will skip the test if the `max_map_count` variable can be
read, and if lower than a given threshold.

Differential Revision: https://reviews.llvm.org/D111465

GitOrigin-RevId: 183f607bd7942ab72c8eff110d15719d63580ef5
Original-Revision: a677a52e2d3491f66c047971a21a4b464237b07d
Change-Id: I9e5aa205e4ea91af4351b4eefcfe70f7f7273aa3