[roll] Roll fuchsia [kernel][vm] Unify Imminent-OOM logic with other levels
Imminent-OOM is a diagnostic level, so we did not trigger eviction at
it. The expectation was that eviction would already have been triggered
before we reached Imminent-OOM, and we synchronously evict anyway at
OOM. However if the pressure level is oscillating between Normal and
Imminent-OOM without hitting the intermediate levels, we will only be
relying on synchronous eviction at OOM, leaving the device at the
Imminent-OOM level for longer.

There can be cases where memory spikes suddenly from Normal to OOM,
leaving us in this degenerate case. Use the Imminent-OOM level as a
trigger for eviction too to catch this case. This will let asynchronous
eviction clear up some memory sooner in extreme scenarios where the
system remains at the Imminent-OOM level.

With this change the Imminent-OOM level is no longer a diagnostic level
for the kernel; it is still a diagnostic level for userspace as
memorypressure clients are not notified.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592024
Original-Revision: b109297c625516753ba2c1c07883d54826f91c48
GitOrigin-RevId: 0d5f4cfb801dc39ad242408ea6c17483533f56d2
Change-Id: Ia9af310acf2a230e678bb6b9ce995fd11ba50877