[roll] Roll fuchsia [kernel][vm] Fix race in ProcessLruQueues
If there are concurrent calls to ProcessLruQueues, then the lru
generation that is read outside of the lock could become stale.
This is a problem since that stale value of the lru would be
re-written to lru_gen_, potentially causing lru_gen_ to move backwards,
which would be an error.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592429
Original-Revision: a51a5062f53982dd955771dff617fd7a86630cd2
GitOrigin-RevId: 263213865b18ca9e48ce9750ffef3490e9df7d66
Change-Id: I86d2150c4fa5fd412c405f31f84da3ad2905d3c2