[roll] Roll fuchsia [wlan][policy] Log MLME event name in disconnect event
This is to prepare for deprecating wlanstack metrics. In wlanstack, we
include the MLME event name in disconnect events in Inspect, but in
Policy, we currently do not. With this change, the MLME event name is
now logged in Policy as well.

Original-Bug: 82654
Original-Fixed: 84892
Test: Refactor + logging change only
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585574
Original-Revision: eb0e71877a74375d03d09a3de173dc7994072e2d
GitOrigin-RevId: 199924410f43513e1e3a81fe938d629f0df16f60
Change-Id: I00f6f843f8bfcdeab91031098b44c326de5e60e8