[roll] Roll fuchsia [botanist] Move timeout context to beginning of execution.
Previously the timeout wrapped the subcommand run by botanist (usually
testrunner). However, before it runs the subcommand, it does some other
steps like booting up the target. Thus, the -timeout flag did not tell
the real timeout of the botanist execution, but rather the timeout of
the subcommand it ran. This would cause a problem when setting the task
timeout because we'd have to account for the time taken by the other
steps in botanist which may not always be enough.

Original-Bug: 67855
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592708
Original-Revision: 1b6e950bfda6fef13062a0599deb90adc40a1461
GitOrigin-RevId: bc4eb20f2f90b1616707ff392daf70ac40cf6b36
Change-Id: Iebf105cd8f19c945a87d2889f3a141c5b1dd32e0