[roll] Roll fuchsia [zircon] Fixed stream read races
There was a race when truncating a VMO whilst reading from a stream
which would result in ZX_ERR_OUT_OF_RANGE errors that shouldn't happen.

This CL addresses this by adding a lock so that they are correctly
sequenced.  This might not be the most performant way to solve this, but
it's simple enough and we probably don't need this case to be
performant.

Original-Fixed: 83803
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589962
Original-Revision: 6ccab18525f2293910339f4b1f3e8b57ed51f8f1
GitOrigin-RevId: df6ddd2617e195b1a28082da22796f67209fbdcb
Change-Id: Ib9c05db9992dfe700e9aed2d1aadfa40975b9219