[roll] Roll fuchsia [ffx] Make self-test cleanup more resiliant
I thought waiting 80ms would be enough to let the daemon exit, this
looks like it is not enough. This instead does a few checks to see if a
daemon can exit on its own, waiting in between checks 50ms. In the end
it then runs pkill to make sure nothing is left running. This *should*
imply that the final kill is only done if something is really getting
leaked.

Original-Bug: 86030
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590564
Original-Revision: 5d8f8f88f29579f48da4b2c1ab72dd822f993142
GitOrigin-RevId: e111fd93c231c1f50ec5ef36107822d552e6a5f7
Change-Id: I68d95ea4d554bb044f883ae6bb60180d02305897