[roll] Roll fuchsia [llcpp] Use generic transports in OutgoingMessage and IncomingMessage
This CL updates the interface of OutgoingMessage and IncomingMessage to
use generic transports rather than channels.

As a consequence, the API is updated so that channel-specific methods
are removed making it less convenient to use these APIs with channels
and requiring the use of an API in ::fidl::internal to read and write to
a channel. An alternative would be to leave the existing APIs on the
classes and route the calls through the generic APIs. However, this will
make less and less sense as more transports are added. Another
alternative would be to move the APIs to create transports from
::fidl::internal to ::fidl.

Original-Bug: 85734
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592405
Original-Revision: afc6bd0cebd2b12fbd37f736b4f2ddaa8f1561f3
GitOrigin-RevId: fb7c0764a49200648598dc524f8c9d8d0b914077
Change-Id: I6c9901f601835a1ee7f6d8607a73de46b827741e