[roll] Roll fuchsia [workstation] Enable new empty account partition in FVM.
This is an initially empty partition with 2 slices, a type GUID of DATA,
and a name of "account". In the future, the user data protection
workstream will reformat and encrypt this partition to encrypt user data
at rest based on the user's password.

Original-Bug: 81606
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/584896
Original-Revision: 429764df081fb6034a5b4cb1d6c9a71eddebe989
GitOrigin-RevId: 0916ea2ffc8420e1858d4bac6c905ae03a3dc8d2
Change-Id: Ibfa96c3ec6e3f8dc93f6ccff368061aa57ee56f8