[roll] Roll fuchsia [llcpp] Simplify unique_ptr/optional<WireSyncClient>
In the early history of `MyProtocol::SyncClient`, the class did not have
a default constructor in order to force the user to provide a channel.
This complicated control flows where the client is lazily initialized:
users either have to wrap the client in a unique_ptr or an optional, and
populate it later.

Today, `WireSyncClient` already provides a default constructor. The
unique_ptr and optional wrappers don't add too much value - the client
itself could be default constructed with an invalid channel, and later
replaced with a valid channel. Furthermore, from
https://groups.google.com/a/fuchsia.dev/g/fidl-dev/c/nF72xpPL4gU we are
planning to switch sync clients to the "->" operator to make calls. This
doesn't interact well with clients wrapped in unique_ptr/optional,
because ones now has to write `(*client)->Foo()` instead of
`client->Foo()`.

This CL explores converting all instances of unique_ptr/optional wrapped
sync clients to default constructed sync clients.

In terms of safety, the new version does not reduce safety, since
neither unique_ptr nor optional check for validity on dereference.
Making a call on an absent client would lead to undefined behavior, and
most uses of unique_ptr/optional are not checking for presence.
Making a call on an invalid client crashes with a consistent message.

In terms of performance, the new version may improve performance since
now there are less boxing and indirections.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592844
Original-Revision: 5c44f0ac7bd77b2a2afd9f43456bcc2b383b9f77
GitOrigin-RevId: 67685500fec74056141010e9c004af4b4858740c
Change-Id: I538ce1a9de52ce72c7a4299f03141a86b5815973