[roll] Roll fuchsia [kernel][hypervisor] Make hypervisor log message provenance clearer
Some log messages emitted by the hypervisor, when read out of context,
have an unclear meaning. For example, the kernel log message:

  Unhandled rdmsr 0x3a

is emitted during the normal boot of a Linux guest. Looking at the logs,
though, it may appear that the host is having problems.

This CL updates log messages in the hypervisor to be clearer that they
are being emitted by the hypervisor, and that the cause of the issue is
the guest:

  hypervisor: Unhandled guest RDMSR 0x3a

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592832
Original-Revision: c304c552d639a42e0300ead040e78b3ce39d6b7a
GitOrigin-RevId: f9f365684d29c35f997af4747ea611e0590fa39c
Change-Id: I6bad2ca90511139c59f93202b231733422ffd15a