[roll] Roll fuchsia [zstd] Optimize unconditionally
Prior to this change zstd was optimized for speed on the host
irrespective of the global optimization settings in order to reduce the
build-time cost of zstd compression. This extends that to the target as
well in order to reduce the resolve-time cost of zstd compression.
Measured in qemu-x64 with KVM this gives a ~70% increase in throughput:

Before:
  $ time fx shell pkgctl resolve fuchsia-pkg://fuchsia.com/debian_guest
  resolving fuchsia-pkg://fuchsia.com/debian_guest with the selectors []
  Error: Failed to resolve fuchsia-pkg://fuchsia.com/debian_guest
  Caused by:
      there is no space available to store the package or metadata
  real    3m53.166s
  user    0m0.871s
  sys     0m0.179s

After:
  $ time fx shell pkgctl resolve fuchsia-pkg://fuchsia.com/debian_guest
  resolving fuchsia-pkg://fuchsia.com/debian_guest with the selectors []
  Error: Failed to resolve fuchsia-pkg://fuchsia.com/debian_guest
  Caused by:
      there is no space available to store the package or metadata
  real    2m18.137s
  user    0m0.817s
  sys     0m0.251s

Original-Bug: 65962
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592869
Original-Revision: 47cf6060ec0900107a1c76a7632e6a73a93ff194
GitOrigin-RevId: 483ee50a8c65686e410c116ebeae0dcd97515a63
Change-Id: If600b28c27883558840d28eb75f2e53d1aec146f