[roll] Roll fuchsia [packet-formats] Refactor TCP header storage
Add the `DataOffsetReservedFlags` type, which stores the Data Offset
field, the reserved zero bits, and the flags. This gives us a simpler
mechanism to copy all of the flags into a new `TcpSegmentBuilder` in the
`TcpSegment::builder` method. It also prepares us for
https://fxrev.dev/I0c8266167ab3bc05885346bdeb6c18b502b93be3, which adds
the [`TcpSegmentRaw::builder`] method.

While we're here, add a getter for the PSH flag, add some documentation
here and in `ipv4` and `ipv6`, and fix a typo in
`ipv4::tests::new_builder`.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/590743
Original-Revision: dbbe288a55f102d1b97f546ca87414e0175cadea
GitOrigin-RevId: b06d51145337e5e3a90b7b27ca05f967ede94fcc
Change-Id: I493cdfca21262b0759870339ba6ce921660f4804