[roll] Roll fuchsia [ffx] Schema from deserialize
Create a library which can construct a serializable witness to a type T
if T: Deserialize. This uses a Deserializer to keep track of the
different parts of T as it is driven by the deserialize implementation
that is implemented/derived for T.

Enums are a bit tricky with this approach, but I worked out a way to
solve those issues. I still need more tests.

The goal is to call this for the output type of a plugin that is
specified as an attribute to the ffx_plugin procedural macro. The result
can then be checked in and compared with subsequent changes to ensure a
consistent schema.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/576565
Original-Revision: 140bc4f2e521b3d3b936fdad04d5419547aec675
GitOrigin-RevId: 81b2f5b834ade004b33605b95c93ae8f934ce004
Change-Id: I0b81fd0421c1ddeeb11a5af924b5dbe26be24908