[roll] Roll fuchsia [wlan][iwlwifi] unbind MAC context in mac_set_channel (if bound)
If an association failed (e.g. AP rejected the client), the driver
could leave the firmware in the some state that crashes the next
association. This is why we saw BINDING_CONTEXT_CMD crash, which
the binding had been set in the firmware.

Since there is no mechanism for MLME to inform the driver that
the association is failed, we check this in the driver. If the
info has been set already, we clean it up before adding new one.

Original-Bug: 86305

Test: Added a test to check if set_channel handles being called
before clear_assoc() is called.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591707
Original-Revision: 6e2bba1c75a63d19f22ddec9e97e5cb19ba29fd3
GitOrigin-RevId: 8978c08f668881c4734a708faddbad8c6ad914ad
Change-Id: Id169273cd59a76871ec56820697885228b82b345