[roll] Roll fuchsia [ffx] Add self-tests for component list
This is basically a smoke test that things more or less work. It brought
out a bug in the self-test runner which is fixed in the ancestor to
this. I plan to add more tests here but this is getting the groundwork
in place.

Original-Bug: 82378

Multiply: ffx-e2e-with-target

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585291
Original-Revision: 732819b51bdf5bf289b61b68f02573d68c55bf57
GitOrigin-RevId: cac9c13aa02c473c50818af98d4663d66e7763c3
Change-Id: I80d56477f935721041f9ad92f9e352d76e2e3a5f