[roll] Roll fuchsia [kernel][x86] Avoid CPU init stack overflows in kasan/lockdep builds.
The x86 arch sets up the kernel stacks for CPU init threads in a special
way, such that the stack pointer is initialized while transitioning into
long mode. Once in long mode, the init thread continues to use the same
stack location for the rest of the init. However, prior to this change,
the stack pointer was initialized to only one page above the stack base,
regardless of configured size for kernel stacks. This can result in a
stack overflow when features such as kasan and lockdep are used
together.

This CL changes the CPU init thread setup to pass the actual top of the
kernel stack to the low-level CPU code, such that the configured stack
size is correctly used and maintained into the higher-level the init.

Original-Fixed: 86361
Test: fx set bringup.x64 --variant=kasan --args 'lock_dep_enabled=true'
boots.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592637
Original-Revision: 3d7113f60f1dfa168652040d896d5a2249975185
GitOrigin-RevId: 9a0b6ff6424c9fe33844f75bcd5e5ec1b8707c9b
Change-Id: Ieccefb7904c9f6f116a27428778b5ac1b036bd5f