[roll] Roll fuchsia [component_manager] Don't auto-bind to ancestors
This change removes behavior that automatically binds to a component's
ancestors when it is bound to. This was meant to support a principle
that parents should run with their children, but this is no longer a
guarantee we aim to provide.

Also, remove duplicate tests in binding.rs (that were rehomed to
lifecycle.rs).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589471
Original-Revision: 5d587d5df5098b5f759f812ddd2af2913dda5934
GitOrigin-RevId: 22600d9348c1bf1ce67472c828a39a8b477c28f4
Change-Id: Ibce5522f47edc71337d66b34f5e4c3aafd4da1d3