[roll] Roll fuchsia [clippy] update script args/usage
Replacing `--files-to-targets` with `--get-outputs` should be more
future proof because when we want to enable clippy targets for host and
other toolchains, we'll need to build a list of output files rather
than a list of targets.

Also introduce the `--all` flag to get the list of, or run clippy on,
every clippy target in the current build configuration.

Also remove some #![deny(warnings)] from the codebase. These annotations
are redundant as we apply it tree wide, and it interferes with clippy
because its lints are produced as warnings that we don't want to deny.

Original-Bug: 86506

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593244
Original-Revision: d140282e8e145581524971cc3c2b65f59910ebcc
GitOrigin-RevId: 0b5be13bd6df1a22e59f1660d2f62e4e5e2cb240
Change-Id: I0d96ac0dc0aa93f4e834b667d0623cfbf5b87ddc