[roll] Roll fuchsia [build][clippy] Add clippy script to action_tracer ignore list
The rust compiler performs some reads of files that aren't truly
dependencies of a given target. For example it will read multiple
libraries in it's library search dirs and pick the correct one to
actually use. This makes it fail our hermeticity checker because those
extra file reads show up as spurious dependencies.

Because we only run the tracer on GN actions, we haven't had to address
this with rustc yet, but clippy (which is a thin layer ontop of rustc) is
implemented as an action, so we need to whitelist it here.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592632
Original-Revision: 2438fe562ac1325386011478fac2490e96bc57e5
GitOrigin-RevId: 5ac50ff7c66647edec28814f48e76fd0b8f1e65d
Change-Id: Id60bca13f30f2b65dd446b1947cea0b24c3eafa2