[roll] Roll fuchsia [bt][gap] Use connection tokens in Peer::BrEdrData
Remove Peer::BrEdrData::SetConnectionState and replace it with the
following methods:

Add Peer::BrEdrData::RegisterConnection method, which registers and
returns a Peer::ConnectionToken. Until the token is destroyed, the Peer
is considered connected. Only one connection is allowed to be
registered at a time.

Add Peer::BrEdrData::RegisterInitializingConnection method, which
registers and returns a Peer::InitializingConnectionToken. Until the
token is destroyed, the Peer is considered initializing (if there are
no ConnectionTokens) or connected (if there are ConnectionTokens).

Instead of explicitly storing the connection state in Peer, calculate it
on demand based on the number of outstanding connection tokens.

Update BrEdr{Request, Connection} to use connection tokens.

Update PeerFuzzer test to fuzz the new Peer::BrEdrData methods.

Rename BrEdrConnection::Start() to OnInterrogationComplete(), which
is clearer and matches LowEnergyConnection::OnInterrogationComplete().

Original-Bug: 70878
Test: fx test bt-host-gap-tests bt-host-fidl-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581303
Original-Revision: 6587e5ff65fece5d56aa6ed0daada5c002015b3e
GitOrigin-RevId: f30bc57f05e35e80e88f59978cebd025d62719dc
Change-Id: I4a608a9afd57afaf7498d62e1b99f3ce78519df8