[roll] Roll fuchsia [plasa] Another internal header in DDK needs exclusion
The added header is not intended for direct inclusion by
the user code, and will error out in that situation.  Add it
to the list of non-public headers so that the Plasa API stub does
not attempt to use it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592603
Original-Revision: eccd73dead7771d00c51727eb4d8bd703e4c3e85
GitOrigin-RevId: 1b0b7711a7e9c395d13eb297ad6174f37cad63a6
Change-Id: I48267b8852e614795c047d7ec15a4e0c1133e47b