[roll] Roll fuchsia [chromium] Allow setting a package name in generated targets
The target name for the browser is `chrome_browser` but the package name
is just `chrome`. Under the assumption that we want to retain the former
for clarity in the build targets, this adds a package_name parameter to
the GN templates. The parameter mimics the existing prebuilt package
templates in //build/packages.

Original-Bug: 86236

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592830
Original-Revision: c18295274dcda7741a0a0cebdf1f0388b7a15b47
GitOrigin-RevId: ea6eac75d2d86ae55b2f8009c5f3a4ba131f8510
Change-Id: I34bca5677280709362b70f9fc831b66e70f701fa