[roll] Roll fuchsia [testrunner] Use a struct to represent flags
Use a local variable struct instead of a collection of global variables.
When all the flags were global variables it was easy to accidentally use
a global variable within a function that was intended to be stateless,
making it harder to test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593703
Original-Revision: 3cd82a11c88d0cb4a1e13cd85c38eac06fb414fa
GitOrigin-RevId: 2f2eb52615c14ed1a76f4be57663c88cd4d03296
Change-Id: I4454ce61fb2cb28dfc3e93461b31fedc1162acd5