[roll] Roll fuchsia [starnix] Add explicit dependency on wayland bridge
Some configurations (e.g., core) do not include wayland (or //src/proc),
but are used for development. Specifically, developers will add
`--with //src/proc` to their fx set. This includes starnix, but not
wayland. This surfaces as an error for a missing capability from the
wayland bridge (child) to starnix (parent).

CQ does not catch this because starnix is not built as part of core, and
on workstation wayland bridge is included.

This fixes the issue by adding an explicit dependency on the wayland
bridge package to starnix.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593365
Original-Revision: df8720afc77ce105a55dff054c05534f015c51c5
GitOrigin-RevId: 0a7f8780b01231866b0e9fb791205a9d250ec8b7
Change-Id: I29acf7005ac8ed62d5ce3efc15a60f3d35a2eac1