[roll] Roll fuchsia [rust][3p] Update trust-dns to v0.21.0-alpha.4
This version includes a change to the default behavior of the resolver,
so that in the common case it does not retry failed DNS queries over
TCP, unless the response was truncated.

Also add integration tests to verify this behavior.

Original-Fixed: 85079
Test: fx test netstack-dns-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594025
Original-Revision: 2da036d581348eb3369dcad831fa7fa8b02dde3c
GitOrigin-RevId: e4cb9aed6cf0f5cec530a4e75d5f9965a3d92ee1
Change-Id: I481f090a21e4e25cd0c4366a675e97917b8966bd