[roll] Roll fuchsia [media/vnext] update fidl files
This CL updates the media/vnext fidl files, primarily by removing the
concept of stream timelines, which seem to be superfluous.

A pixel format modifier has also been added to the video format
representation. This is a placeholder and needs to be fleshed out to
have the same semantics as are found in sysmem.

Test: fidl files only
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594537
Original-Revision: 769df1b22d03c5bdbacac47f20a8bf09092c3e3f
GitOrigin-RevId: ff4f0d034be25839cd5eda610f041013001327ec
Change-Id: I0b7fe4d5d3cc357d53b0bd7767c43682f9ef6f46