[roll] Roll fuchsia [kernel][vm] Process the DontNeed queue
Process the DontNeed pager queue as well when processing LRU for page
queue rotation. This ensures that any pages in the DontNeed queue which
have been subsequently accessed, are fixed up and moved to the
corresponding pager queue. If the DontNeed queue is not processed when
advancing LRU, the queue indices in the vm_page_t can become old to the
extent of appearing to be valid again once the queues wrap, throwing off
queue counts. Currently the DontNeed queue is processed only when trying
to evict from it, which is not sufficient.

This CL introduces a pair of toggle DontNeed queues, and a generation
count to track which one is the current DontNeed queue. During
processing of the current queue, pages that do not require fixup are
moved to the toggle queue. This allows the processing thread to drop the
PageQueues lock between performing fixed amounts of work, and resume at
the tail each time until the current queue is empty, without having to
maintain a cursor which would be tricky with dropping the lock.

This fixes an existing bug where pages in the DontNeed queue which have
been accessed can result in incorrect active/inactive counts.

Test: kernel-unittests-test

Original-Bug: 85768
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592153
Original-Revision: 495e5648edb42c5da3c11a4918502cb49632cd41
GitOrigin-RevId: d7bd0776fcbbea98ced4fc07821ff4714913c265
Change-Id: Id0116c7522bc4f2ea798045ece97a17d267be574