[roll] Roll fuchsia [ffx] Fix e2e test flakes
I added this code to kill zombie daemons but it is too aggressive and is
causing flakes. This makes it not an error to kill a left over daemon
but still does the kill if necessary.

I also increased the per-test timeout to 12s because 10s is a race with
the default proxy connect timeout so the error messages are test case
timeouts usually instead of the real timeout that is causing the
problem. This should make those flakes provide more diagnostic
information.

Original-Bug: 84938
Original-Fixed: 86030

Multiply: ffx-e2e
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594862
Original-Revision: 2ef4d8dfc6a308f5986f1fc992cb55e7f58b40ce
GitOrigin-RevId: a274f12aee50b84a1e06394fee3477b2f8c2c837
Change-Id: I932d16964cabdf61b98bcbb73b3a0ea681e29a71