[roll] Roll fuchsia [ffx] Add tests and cleanup target formatter
This adds some additional test coverage to the target list formatting
code. I also used this as an opportunity to change some of the patterns
in the code to be slightly more compact and I hope more readable. Some
of it should actually be more efficient. The tests were added first and
then the refactors to ensure they are true refactors.

Original-Bug: 86588

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594021
Original-Revision: 4b5663569ed4d072e9c07484ee8196be58cacc2a
GitOrigin-RevId: eca36bb4c513e16666e4853dd4565fb8dab456da
Change-Id: If7baa3292fe3838ef8114d9e68c3ac2dc5973ac4