[roll] Roll fuchsia [magma] Rename RegisterIo::Write32 to Write32Flipped
This is part 1 of flipping the order of the arguments to
RegisterIo::Write32. The order of the arguments to the RegisterIo::Hook
methods is also swapped for consistency.

This patch makes it easy to identify which callers are using the old
ordering. Part 2 will reorder the arguments everywhere and rename
everything back to the old names.

Original-Bug: 86716
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594864
Original-Revision: 9f487ec58a359af51b00e11bbfec5ca23e4475ca
GitOrigin-RevId: d22a7312391d2ef874f3139436a4fa054eef4557
Change-Id: I4a514e7205c09125406f8845f1bbca104ef17d4d