[roll] Roll fuchsia [driver_manager] Start separating v1 code into a subdir
This change reorganizes code that is only used in DFv1. This v1 code
is largely in maintenance mode as we plan to flip over to v2 in the
future. Future changes will continue to split out existing classes to
make it easier for us to delete this code once we no longer need it.

Header includes were also updated to be style guide compliant.

Test: No functional changes
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594748
Original-Revision: a596e3c110511b6ad4e26366bd29b4cb5163c0aa
GitOrigin-RevId: b94adf8d03fdbdea2014b22e91616f1fbbddb6f6
Change-Id: I59276ac3bcf3dea587e44ffa1437d21f66e6e7b2