[roll] Roll fuchsia [llcpp] Specify wire format version for device metadata
device_get_metadata returns a buffer containing FIDL bytes.
Currently, the wire format version for these bytes isn't
specified on decode. When the version is not specified, it
is assumed the initial bytes contain the FIDL header which
can be used to determine the version, but this is not the
case. This CL adds explicit naming of the wire format version.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594081
Original-Revision: 27f44ff662b79644b8379c6df011720dcdc44b85
GitOrigin-RevId: 7b96446f470704e5643faf792d72a33901a5a0df
Change-Id: Iae79b8ae7447c90fa66fd4f7bdff80585a050f28