[roll] Roll fuchsia [ffx] Nest emu-remote as a subcommand of emulator
Part of removing the experimental flag from `ffx emu` is dealing with
the usage of `fx build` that's part of the emu-remote workflow. Rather
than modifying the workflow, which isn't stable yet anyway, we decided
to move the remote functionality from `ffx emu` and put it behind its
own experimental flag which can be dealt with when that plugin is more
mature.

This CL performs a fairly significant refactor on the `emulator` code to
nest all of the subcommands as child plugins. The `start` and `kill`
subcommands share the original experimental flag, while the remote
subcommand has its own new flag. Most of the command logic is moved to a
`common` library subdirectory, so that all of the emulator plugins can
share it.

This CL does not alter any functionality, it's purely organizational.
The move does alter the build targets for the plugins and their tests,
but all of those tests are still valid.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594282
Original-Revision: f3ef1eb5f633becd918393b86bdc66a25e4fcbff
GitOrigin-RevId: 5b9dd51ce432b904f81e8901cb5d3fa16a493192
Change-Id: Id0df0b9d8aaa4930c1bc41723c826e88dbfc8077