[roll] Roll fuchsia [ffx] Add golden tests for target list JSON
This change is to introduce tests that did not yet exist. There is no
functional change here. There is a bug in the json format that needs to
be fixed, but I wanted to get these tests of the current behavior in
place first and then update things in a follow-up.

Original-Bug: 86588

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594022
Original-Revision: d7ebf4bbeb06d2039a9ab4891df4a770353f6abf
GitOrigin-RevId: af493f1dfeacf9e3c31e5c518d70a57d9fb93c1f
Change-Id: Ie48e19674b73496c259727f39d9dc223c694c048