[roll] Roll fuchsia [build] Create drivers global documentation file
Generates a global documentation file with driver provided
data. This data is optionally included in a JSON file in driver's
folders. The generated global files is a JSON file to be used for
display on fuchsia.dev.

Deployment of the generated global file, validation the driver
provided JSON, and documentation are to be provided in subsequent
CLs prior to deployment.

A single driver provided data file is included for testing.

Original-Bug: 81799

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593684
Original-Revision: f19984b474fc7cfd412367f33bd60dcbfd59ba5f
GitOrigin-RevId: 9684eb979bd8299aa7e66c8f2bba933c5098692c
Change-Id: I4f1c53ec444a8a3340a45e63c564c0a4758dfba1