[roll] Roll fuchsia [ffx] Add correct target_type to target list JSON
The JSON formatting code and the tabular formatting code use different
logic for computing the target_type field. I think this was just an
oversight when the tabular format was updated to take into account the
build configuration information, as the JSON code was not updated at
that time.

This unifies the logic for both so hopefully it is clear now if
something changes for one that it will/should change for the other.

Original-Bug: 86588

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594023
Original-Revision: 72117540255f3f445e894494f81782dbaff94dc2
GitOrigin-RevId: eeb76ec2f1230d11cd06558af0e419054eadeb72
Change-Id: Idede612187415dd7b615637d2677c27dc0b47c0d