[roll] Roll fuchsia [driver_runtime] Add handle implementation.
This will be used for implementing fdf_channel_t.

This is closely based on the kernel handle table implementation.

Each process supports up to 64K handles. This is implemented
as a handles table directory which can hold 64 tables,
of which each table holds 1024 handles. We allocate 1 table at a time.

The fdf_handle_t value consists of:
 - the generation number: incremented each time the handle is reused
 - the directory index number: for locating the handle table
 - the index number: for locating the Handle in the handle table
 - reserved bits: 2 LSB, where LSB is 0. This differentiates it from a
   zircon handle which has both bits set to 1.

The directory index and index number allow for quick lookup of a
Handle given a fdf_handle_t.

Deleted handles are recycled onto free_handles_list_, which
can be used for new allocations. When reusing a handle, the
same directory index and index are preserved, (so that we
can correctly locate the corresponding Handle object),
but a new generation number is set to avoid returning a
recently deleted fdf_handle_t value.

Original-Bug: 85577

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593369
Original-Revision: 3afc34a553fa3c49c925cbbb52663de5f5bad66f
GitOrigin-RevId: 3155cde9a3db23cd217b00a05535e1ba0a8d2d5e
Change-Id: I5b2e14dc0a52257c0bd404de2210f8c0bc1ae5a5