[roll] Roll fuchsia [a11y] Return status from input.accessibility.PointerEventRegistry
This change modifies the semantics of
fuchsia.ui.input.accessibility.PointerEventRegistry to match
fuchisa.ui.policy.accessibility.PointerEventRegistry so that we can
later replace the policy namespace.

NOTE: This is NOT a breaking change. Root presenter was the only
implementer of this protocol, and we removed the implementation in
this change: https://fuchsia-review.googlesource.com/c/fuchsia/+/496699.

TEST: n/a
Original-Bug: 86102
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592601
Original-Revision: efe120b7d47e19f440ff1d2cb67f821fec48ae40
GitOrigin-RevId: 6d441f05658551c7a03e77a71d199d5482396de9
Change-Id: Icf0c98caddee30b2d09a049b59f904a7cdc48d18