[roll] Roll fuchsia [starnix] Add wayland server implementation
This adds a wayland server that is spawned when a component adds
`features: [ "wayland" ]` to their component manifest.

The server connects to wayland bridge in a couple ways:

  1. A channel is provided to the wayland bridge that is used to
  communicate wayland protocol messages from the wayland client.
  2. Starnix connects to wayland bridge's ViewProducer, which will
  provide a ViewProvider for the wayland client.

The server also communicates with the session by serving the
ViewProvider protocol on behalf of the component, so it can be added to
a session as an element.

Finally, the server communicates with the wayland client via a socket
and "memory allocation file." The server allocates shared memory on
behalf of the wayland client, and creates file descriptors for the
associated files.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594146
Original-Revision: aa6a9a6326e0d3dad322f50134b42a6092e97e99
GitOrigin-RevId: 0c1ac5e5f5a27eb990aa9177a73f9b20825cffcc
Change-Id: Ie518ad4b113041796dfb1067c661579c0ef60f42