[roll] Roll fuchsia [blobfs] Use OP_DONT_NEED for inactive blobs
When blobs transition to being inactive, hint on their associated VMOs
that their pages can be evicted under memory pressure. Currently, this
is done under the covers with a kernel hack that performs the same
operation when no clones remain for a VMO, which is when blobfs
considers a blob inactive. The kernel hack is going away now that the
proper hinting API is available for use.

Original-Bug: 65334
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594978
Original-Revision: 58399e8cdfdd3773daedace038ff5a4683320f27
GitOrigin-RevId: 2ff1839892a2c7e24eac53396586f82823232c85
Change-Id: I7a946cecadc5b065b8a0bc2373ceea94618354b2