[roll] Roll fuchsia [fx] Add ffx device discovery to `fx shell`
This is a fix-forward of fxr/593784. It includes the changes originally
contained in fxr/591990 and fxr/592608. The original change didn't work
with the fx flag "-d", used for specifying a device by name on the
command line. The change from the latter CL fixed that issue.

Original change descriptions:
591990:
fxb/85247 describes a shortcoming in the ffx emu workflow, where certain
fx tools couldn't communicate with the emulator instance unless the user
manually indicated the address and port of the target device using `fx
set-device`. We tracked this down to the fact that the `fx shell` tool
calls a library function which only checks environment variables and
hidden files for device information, and doesn't check the ffx daemon.

This CL introduces a new library function to the fx "vars" file,
get-fuchsia-device-port. This function starts by calling the original
ssh port function, get-device-ssh-port, then falls back to ffx if it
doesn't find one. We then modify `fx shell` to use this function instead
of the original, which waterfalls to other tools that call shell
directly (such as debug and fidlcat). This removes the requirement for
the user to call `fx set-device` for these tools, even when running on
SLIRP instead of TUN/TAP.

There are about half a dozen other tools which still call the old
function directly, and these can be migrated in future changes if deemed
appropriate.

592608:
A regression was introduced by the use of `fx-command-run host-tool ffx`
that does not pass fx and build configuration information to the ffx
tool as is typically expected within devshell tools. A new function
simply named `ffx` is introduced that provides an easy and correct way
to launch ffx from fx subtools.

Original-Bug: 85247
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594845
Original-Revision: 1cb54a51bd129e642ef35b5fbb9199205178c495
GitOrigin-RevId: e89a7be1f50de6918307422bc2c3c96a6144ee90
Change-Id: I58fa2abcbae56fb8f8c72f36d796f4630500ddc7