[roll] Roll fuchsia [build][kernel] Simplify specifying kernel command line args
In the current build system, if a developer wants to specify an ad-hoc
kernel command line argument for local testing purpose, they have to:

1. Create a new directory, `local`, in their Fuchsia root directory.

2. Create a new file in this directory, `BUILD.gn` which contains
   a valid `kernel_cmdline` target, in turn containing a list of
   strings to add to the kernel command line.

3. Update the `dev_bootfs_labels` argument in their `args.gn` file
   to point to this new rule.

If the user has multiple builds and wants different sets of kernel
command line parameters for each build, the `local/BUILD.gn` file needs
to contain multiple different kernel_cmdline rules: one for each target.

This makes it hard to tweak kernel command line parameters for the
purposes of local testing and debugging.

This CL a new GN argument, `dev_kernel_cmdline`, which allows a
kernel command line argument to be specified by just adding it as
a string:

  fx set ... --args='dev_kernel_cmdline=["my_option=value"]'

Values in `dev_kernel_cmdline` are not expected to be checked in, but
rather are for local development only.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594705
Original-Revision: e9bf9dfccce39d3ce9fe0b22f5c48d094b5a7840
GitOrigin-RevId: d645902168f5bcc9bb2217107325e41a0153c478
Change-Id: I0c6eb6237cbf9fb928018d963127a676abb5a20a