[roll] Roll fuchsia [fidlc] Add ForEachDeclMember
This CL adds a generic facility for applying a function to each
Attributable* member of a Decl*, and rewrites VerifyDeclAttributes in
terms of it. This reduces duplication and the chance of mistakes (in
fact there was a mistake: table members were validating placement twice
instead of placement and constraints). I also plan to reuse this for
FIDL versioning.

Test: fx test fidl-compiler
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593744
Original-Revision: 59562966bb1dc2a8b0137c4e365cfccd4fbf378f
GitOrigin-RevId: a81b98dec03be5139176f7f5679efe15d845eb14
Change-Id: I584fd6bda128aa5cf0a60fc85ef648d6028f53d5