[roll] Roll fuchsia [fx] Add ffx device discovery to `fx shell`
This is a fix-forward of fxr/595456. It includes the changes originally
contained in fxr/591990 and fxr/592608, plus an update to `fx
serve-updates`. The original change didn't work with `fx serve`, because
that script used a call to `get-device-ssh-port` directly, as well as a
call to `fx shell`. The discrepancy between the two results led to the
`fx serve` script looping over the call to `fx shell` repeatedly,
waiting for a connection and discarding the error message it was
returning indicating it couldn't reach the target.

This CL replaces the call in serve-updates with one to the new
`get-fuchsia-device-port`, so the results in `serve-updates` are always
consistent.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594885
Original-Revision: 5554db46c78211663f00333c334d55150bf69755
GitOrigin-RevId: ef0eafce3f456f813a9e29571a551d6baf23623a
Change-Id: I48157bde2b647ee00bbd1ad3d230eecc1d97cc32