[roll] Roll fuchsia [a11y] Modify "scene ready" condition in semantics tests.
The web semantics tests hang during setup waiting for a
ViewStateChanged event indicating that the web view is rendering. Scenic
sends this event on the first render pass that visits the web view after
it's been attached to the scene. It's possible that there's some
non-determinism about when this render pass will happen given that we're
not immediately loading any web content into the view (in which case we
may not request a draw operation). Since we really only care if the
web view is attached to the scene at this point (and not that it's
actually rendering content), we can instead wait until the test fixture
view is attached to the scene and the web view is connected to its
view holder.

TEST: ran flutter and web semantics tests 100 times each.
Original-Bug: 86243
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595385
Original-Revision: 7ca8b3c7aaa62f94a77f8688238e71673d64795b
GitOrigin-RevId: 1f1482c95051f4352473d78ff76c28a29d50eb08
Change-Id: I8d6684f29084358842c1a9031b5ec6f23359655d