[roll] Roll experiences [Ermine][Quicksettings] Add initial WiFi UI
Adds initial WiFi quicksetting UI to view scanned networks. As a temporary solution quicksetting is disabled until the scanned networks list is populated. This has two purposes: the first one is to prevent the quicksetting appearing active before scanning is available on device, as well as prevents the user from accessing the wifi quicksetting until after the initial scan has taken place (takes around 20 seconds initially). Networks that are not compatible are disabled/greyed out. Note that there is a bug where network names with non-ASCII characters may appear in the scanned networks list and appear as blank networks in the UI. This will be addressed in a subsequent CL.

Original-Bug: 79885

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/592913
Original-Revision: 0aac4c9cbacb45cfe749be8953791fa435146b3a
GitOrigin-RevId: ce39d23e09aa1613ab6be1048d0268437cb55a06
Change-Id: Ic30127d18ee24f29bd517793858701b14ae33dd8