[roll] Roll fuchsia [fuchsia-async] socket2::Socket to DatagramSocket
Add a method to construct a DatagramSocket from a socket2::Socket. This
is necessary for use cases which need to construct a socket directly
via the fuchsia.posix.socket FIDL APIs. The function checks to ensure
that the socket is of type DGRAM, and sets it to be non-blocking.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594976
Original-Revision: d5076b4b57218b084336eaf2bd2f65c2a6cfa1df
GitOrigin-RevId: b22a8606260b9840bcf48658abb2a649505dd209
Change-Id: I9acb579df4bad465d62ddcb1f8adaa64b46fda06