[roll] Roll fuchsia [blackout] fix tests for local running
The integration between the host and target code of blackout was not
tested, so the mechanisms used to run the target code on a device
rotted. This fixes the rot, now blackout runs again on local devices. A
followup commit will add the necessary things to make it run in CI.

I'm not including a test here for the host-target interaction because
the current strategy is deprecated and fragile, and I'm investigating a
more robust strategy, but I want to get the fixes in because the higher
priority is to get this running again.

Original-Bug: 44392
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/571550
Original-Revision: 6f353cb4b3adb388e3e42faedb1e46fde384f466
GitOrigin-RevId: 9a3c9f9f97c48408cea8367389f02c5eb0184dc1
Change-Id: I8201e202553f841cf179903050b420ca47c33d9c