[roll] Roll fuchsia [libc] Add GWP-ASan note in libc.so
A new ELF note ".gwp-asan-info" is added to the libc.so which could be
used by out-of-process crash handlers to locate the GWP-ASan state. The
note contains an relative offset to __libc_gwp_asan_info, which is
initialized to contain 2 pointers to the allocator state and the
allocator metadata during libc startup.

Crash handlers are expected to use the functions and data structures
provided in //zircon/system/ulib/c/scudo:gwp-asan-common.

An integration test for GWP-ASan is also added.

Original-Fixed: 78949
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593054
Original-Revision: b14ff258353aa9a853a703d99955e57b3a20de88
GitOrigin-RevId: 2abdd43270a49a70e8938dddb48a97d4e4dadf19
Change-Id: I86d0a7bdea9dcd69b8c7a427aa24541108d1ab87