[roll] Roll fuchsia [rust][fidl] Remove redundant uses of .check()
This CL removes calls to check() when they are immediately followed by
an await, because this is redundant. The check() method for two-way FIDL
methods only exists to check the status of writing the request to the
channel before awaiting the response.

Original-Bug: 74371

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595245
Original-Revision: 14ac3e060e3b0146bd7dac0328e1115d0f9b464f
GitOrigin-RevId: 70799da5ba76591dc3e457eb6f7b259335dfe262
Change-Id: I84593e77d0e628bc02cef8ade432bfae58dea4c3