[roll] Roll fuchsia [brcmfmac] Make all devices have the same parent
Currently brcmfmac creates a root device that is bindable, binding to
the wlanphy driver, then it adds children to that device. This
potentially causes problem for device manager because child devices are
now binding during binding of parent devices.

This change adds the children to the same parent that the first wlanphy
device was added to. In practice this makes the wlanif devices siblings
of the wlanphy device instead of children. This allows the wlanphy
device to safely bind to the wlanphy driver.

Original-Bug: 86567
Test: brcmfac_test, brcmfmac_sim_test, manual testing
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594750
Original-Revision: e6af89a83d8a0732beaff8895328cf07513bd054
GitOrigin-RevId: 4d82ae24200715dc0baba50c203ac2f26b4d6be9
Change-Id: I42d435b392739faaac58b1418197303f6e77d415