[roll] Roll fuchsia [bootserver] Retry downloading images on CRC failures
We've been seeing a fair amount of checksum failures when downloading
images lately, and it seems to be transient (doesn't happen consistently
for a given blob) and not necessarily associated with the Go GCS client,
as other teams have seen similar failures with other GCS clients.

The official GCS docs say it's okay to retry a download in case of a CRC
failure, so this is probably fine to have anyway:
https://cloud.google.com/storage/docs/hashes-etags#validation

Unfortunately the storage library doesn't export the bad CRC error:
https://github.com/googleapis/google-cloud-go/blob/07c804b08b9f2bbe181ffcee1b9b41005463b057/storage/reader.go#L415
So we must resort to a substring check to see if a download error is
caused by a CRC failure.

Also apply retries to corruption errors from the `flate` library, which
are suspected to have a similar root cause.

Original-Bug: 70972
Original-Bug: 84619
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595882
Original-Revision: 33c23fc009c289fe97fb303a2b68bdb969e02b9c
GitOrigin-RevId: 97aaf757eede7b4b310f41fab6168feedebee227
Change-Id: Id5dd114c73aa3b5014ab1de99bdb4d03604623f6