[roll] Roll fuchsia [audio][intel-hda] Fix race in completer replies
The completers and any member of StreamChannel objects need to be
protected by obj_lock_ given that the stream_channel_ is protected.
Moved the locking to the call forwarding done when StreamChannel
addresses are taken and added a comment to make this clear.

This was uncovered as a flake.

Original-Fixed: 85445

Test: fx test intel-hda-stream-test
Multiply: intel-hda-stream-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596066
Original-Revision: 92683d2fc1e1c981f92729cd334a43fbfb18741a
GitOrigin-RevId: d3a963bdcde50f58ed4e97aa4970bf6ea8e7355c
Change-Id: I4f89282f8b436f227eb2e5ff1184698d6205a008