[roll] Roll fuchsia [usb][xhci] check device state
Originally, we were getting transfer events after a device was detached,
which could happen because ordering of events is not guaranteed. By
accessing the array through the index (GetDeviceState()[i]), we
weren't checking whether or not the device was still connected. Thus
causing a page fault when trying to get the address to the ring.

Now, by accessing the array through the method (GetDeviceState(i)), the
the method checks if the device is connected and returns null if it
isn't. We check if the returned state is null and abort the transfer
event when the device is already disconnected.

Original-Bug:86967

Test: Tested on Atlas

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596061
Original-Revision: 6f898137dfbc400e1615936a6789fda08e7c6f08
GitOrigin-RevId: 21330d998af7218982817fd6466d97bdadc415f7
Change-Id: I2b2203ff02e52e5d42719816018af3964e8f71df