[roll] Roll fuchsia [fidl] Remove option bits from fuchsia.fs.GetInfo.
These make it difficult to call and difficult to implement. The
performance of computing the flags is trivial except for the device path
which should itself be relatively quick. We do not forsee adding
additional difficult-to-compute data to this structure.

This is more important as part of bug 84558 where we want
DirectoryAdmin.QueryFilesystem to have the same behavior. That is
called/implemented in more places so these bits become more
inconvenient.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/589494
Original-Revision: 5dd8324419be77f36991075f03c7f598de19fd5c
GitOrigin-RevId: f630207471eec62cb09dc2a2aa56dc864a68f01c
Change-Id: I68364b915e240d73781b0ae23063c8b93781dc7c