[roll] Roll fuchsia [media/vnext] add StreamQueue to vnext/lib
This CL is part of a prototype for media vnext. StreamQueue is a
thread-safe queue whose elements are packets, end-of-stream
indications and clear requests.

StreamQueue is part of vnext/lib/stream_sink, which will also contain
implementations of both ends of fuchsia::media2::StreamSink, to be
added in future CLs.

Test: unit test provided
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595914
Original-Revision: 126a27cb6576ed0ce31ad816869ec46ed93ce453
GitOrigin-RevId: a9b6bcdd257282f2543e201e9489c8be8525ce2f
Change-Id: I8a29ba47b49f1c03b960bc24610e2d65560ba371