[roll] Roll fuchsia [rust][rbe] Include .so with .rlib deps.
Include .so with .rlib (for those crates that have them), to get more
consistent results between local and remote building.
The depfile generation revealed that the .so was missing in some cases,
which caused compilation/linking to fall-back to using the .rlib, when
the .so was actually intended.

https://github.com/rust-lang/rust/issues/90106

Original-Bug: 86896
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596429
Original-Revision: 076e690a1b83f19fdae4b260daade555e3bb1cb5
GitOrigin-RevId: f6601af7613ef6474e306e97b280eb489c77cb44
Change-Id: I814630ce04e25f9db101c60fef80a299d400ab10