[roll] Roll fuchsia [bt][mock-piconet-server] Clean up request handling
Background: The main task of the server selects over a number of futures
while handling FIDL requests. Upon reception of a MockPeer request, it
invokes a handler which asynchronously sends a message. The receiving
end of the message is also processed by the aforementioned main task,
and handling of said request is blocked on sending the message. Our
theory is that this could have resulted in a sort of circular
dependency, causing flakes due to stalls.

Instead of asynchronously sending the result of the handler, we can
simply return the result synchronously.

While here, improve logs.

Original-Bug: 82484
Test: fx test mock-piconet-server-tests
Test: fx test bt-rfcomm-integration-tests (which rely on the Mock
Piconet Server)

Multiply: bt-rfcomm-integration-tests

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593056
Original-Revision: 233ecd2987d1293ca1bdd11268aa869504345e3a
GitOrigin-RevId: b6f38d530ce3a3030bf64773821a66ee13d1931e
Change-Id: Ie591f8790d06b97f911d2a20a94c42e2ee4d4b3a