[roll] Roll experiences [workstation] Attempt to get an account directory on launch.
With the current state of the backend implementation, this reaches the
call to get account with fails with UNSUPPORTED_OPERATION, this (and any
other account-related failures) is logged but is not fatal.

I reason that there is little value in adding unit testing to
workstation session for this feature we will be moving into login shell
within a few weeks, but I can do that if necessary.

Original-Bug: 81601
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/596761
Original-Revision: 8dc5b308c5dd75ec1230b1d808894bf2a8440d12
GitOrigin-RevId: 26edd0a24660ca0c3eb66864231657bde70d9557
Change-Id: I4f917f03710a831abae83996a8cebe21a13bdd4a