[roll] Roll fuchsia [vfs] Move QueryFilesystem to FuchsiaVfs.
This is designed to make it easier to implement QueryFilesystem on all
filesystems and to unify on one implementation. Currently this is
exposed in different ways.

With this change, a filesystem needs only to implement GetFilesystemInfo
once on their VFS and it will work for all nodes, and they need only to
instantiate route the fs::QueryService to implement that API using the
same implementation.

Adds more Vfs backpointers to the root nodes in the Blobfs root.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596749
Original-Revision: 9dcadba26335fbce86fe9d5af39a9c9053a7d32e
GitOrigin-RevId: 40e9fa2e3686281544aca14aaaf7db3088f2d2d4
Change-Id: Id42eaf1f090c7cd8f2a2bcdb85ccd2b9a48cbbf0