[roll] Roll experiences [ermine] Remove peek from right edge to allow scrollbar access
Ermine allows displaying the sysui using pointer peek on the left and
right edge of the screen when apps are fullscreen. But this interferes
with the user trying to use the scrollbar on apps like Chromium that are
also on the right edge of the screen.

We still keep the peek behavior on left edge, as well as the keyboard
shortcut (Meta key) to display sysui when apps are fullscreen.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/597147
Original-Revision: 357ab78f6da02af0e7a150e88140844db693e8dc
GitOrigin-RevId: c8a6de135e7ce021bd9636381b60b3d035659bdc
Change-Id: I838eafe51aaaab65dc4e136fb55bc0ec94f631cc