[roll] Roll experiences [e2e][simple-browser] Converted ermine_testserver to a flutter component
ermine_testserver was launched using Component.launch in e2e test and therefore, it did not have any views. However, closing the test server through the workaround using a special URL turned out to be very flaky in testing environment. For this reason, this CL converts the ermine_testserver component to a flutter component with a view so that it can be closed by closing its view. Verified that the test was much more stable and passed ten times in a row on local machine.

Original-Bug: 80345

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/596406
Original-Revision: 80ca85648970df9a97003d558b63efa6a16b3895
GitOrigin-RevId: 4eada3dffff0641ec0e2362618104bcbf72ab89c
Change-Id: I63b5b142083f8fdced7a1949445bfe0cab29ce17