[roll] Roll fuchsia [inspector] Support for GWP-ASan exceptions
inspector_print_debug_info will now print the GWP-ASan information if
the thread is on exception and the exception is caused by GWP-ASan. The
information contains the error type and the allocation stack, as well as
the free stack if the allocation is freed.

Test: in a Fuchsia shell, `export SCUDO_OPTIONS=GWP_ASAN_Enabled=true:\
      GWP_ASAN_SampleRate=1:GWP_ASAN_MaxSimultaneousAllocations=512` and
      run `crasher use_after_free`. The use-after-free is captured by
      GWP-ASan and the crashsvc prints extra info.

Original-Fixed: 77557
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595642
Original-Revision: 9cde8598fc91578f24e297085c5ecabc6d20e65b
GitOrigin-RevId: 14fd87a014166676dcc7557e415a6a425c3c7f44
Change-Id: I1607bf73a1f6f73831d518f98554ba747050e6ba