[roll] Roll fuchsia [media/vnext] add the rest of vnext/lib/stream_sink
This CL adds the remainder of the stream_sink lib, stream_queue
having been added previously. Included are a StreamSink client and a
StreamSink service implementation. The former pulls from a
StreamQueue while the latter pushes to one. This pipeline effectively
proxies a stream queue across a process boundary.

In addition to the client a service implementation, this CL includes
a few files that define associated mechanisms, including a fidl-
specific clear request, a fidl-specific release fence for packets,
and a template for conversion between fidl and internal packet
representations. These are tested along with the client and
service implementation in the unit tests.

Test: new unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597182
Original-Revision: 8e76569fd226e6cb5f581e0f80c779349e63745b
GitOrigin-RevId: 60aca1cc5d98ff33b0f16fb53106186dd943a511
Change-Id: I8e8b49de34ca6fd174ad3e9ea7c094fe6169772a