[roll] Roll fuchsia [kernel][spinlock] Check that ints are disabled in IsHeld().
It is invalid to call SpinLock::IsHeld() with interrupts enabled,
due to the use of the CPU number to indicate the holder.

Add a debug assert that interrupts are disabled when calling
IsHeld() to avoid recurring mistakes using this method in the
wrong state.

- Add debug assert for ints disabled to IsHeld().
- Simplify code that asserts ints are disabled + IsHeld().
- Fix call to IsHeld() that may occur without ints disabled.

Test: Updated test does not assert.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/460772
Original-Revision: f1deee8ac2fdbabda0af3b267299b2ec7d47ed2d
GitOrigin-RevId: 75b8dd9973d037b08e315e55e6c99bc329ddcfb5
Change-Id: Ia47c2a132e1588c0e33062eb1ff0dcea2c135a05