[roll] Roll fuchsia [audio_core] Add per-pipeline-stage statistics
To help debug underflows in the wild, add per-pipeline-stage statistics
that are derived from zx_runtime_info_t. If a pipeline stage reports
stats, it should be given a unique name. A future CL will assign names
in a more careful, meaningful way.

For now, we report statistics only from EffectsStageV2 nodes. These
stats are reported directly by the out-of-process effect via the
response to fuchsia.audio.effects.Processor.Process.

This CL changes fuchsia.audio.effects.Processor in a
not-backwards-compatible way. This is safe because that type is
currently not used out-of-tree. (After this CL is submitted, the type
will be used out-of-tree, so this is our last chance to make changes
without a LSC.)

A follow-up CL will move the file static_vector.h into the fbl library.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592679
Original-Revision: b4c10a3abf01488a8dfd38359b5f30d12873e335
GitOrigin-RevId: a3c43b204947752ebdbf68bb09526faf23601166
Change-Id: I063af5e89f125f15b685409970fbf50989f27606