[roll] Roll fuchsia [factoryfs] Update to use new Query interface.
There is a new QueryFilesystem interface in the VFS layer that
simplifies some of the implementation. Use this. The custom QueryService
in factoryfs can now also be replaced with the VFS implementation.

To enable this, the VFS pointer needed to be plumbed into the Factoryfs
class which meant updating the constructors. As part of this, the
::Create functions now use the new zx::status syntax. The initialization
now matches Blobfs again which is what this code was originally based
upon.

Remove an unused function in Blobfs I noticed.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598114
Original-Revision: c3ae8a5926abc836b619c00984753a2423c78645
GitOrigin-RevId: f29df0afe17041f5376c9d28eab0ec827d725b19
Change-Id: I9043d6efe45de711a942db13739ced4a1c233332