[roll] Roll fuchsia [starnix] query_events
In FileOps::wait_async, we check the current events
and immediately queue a packet if the the events that
we are interested in are present. We need to do this
so that waiting on Zircon and non-Zircon file objects
have the same behavior. All Zircon kernel waits behave
this way.

The achieve this, we need query_events to check the
current event state of the FileObject.

Original-Bug: 86403

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/594965
Original-Revision: 8978990e27144c81537f4247632e5b5fb208002a
GitOrigin-RevId: b6d902be5075f820ab234872eaa4609b27ae0740
Change-Id: If496aa7fee569a32218bd2ac876c68c2100aa3b7