[roll] Roll fuchsia [bt][a2dp] Fix leaks in Permits, cleanup
With revokable permits, a revokation function could hold a
reference to the Permits. This would cause a reference loop to the
PermmitsInner when dropping the permits, meaning it would leak if
everything was dropped at once.

Have the Permit and the Reservation hold the Revokation function
themselves, so if they are dropped, they will drop their own references
to the Permits.

Some other cleanup as well:
 - Cleanup unused results lints and turn on linter.
 - Add some logs on unexpected results.
 - Switch from log to tracing.
 - Switch to fuchsia::test where reasonable for logs in tests.

Original-Bug: 75375, 86726, 81181
Test: updated, fx test bt-a2dp-lib-tests
Test: run LSAN and confirm no leaks
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/569544
Original-Revision: 28ae66140a1f862cd4687fba4d28b9395330340f
GitOrigin-RevId: 2616004f8dafb8ab5d236f4391170aa97da8a7c3
Change-Id: I8c8c4eab0d68943151a113e996b1cf325f738387