[roll] Roll fuchsia [media/vnext] add vnext/lib/threads
'threads' is a small library that implements the fmlib::Thread class.
fmlib::Thread is essentially a wrapper for a shared pointer to an
async loop and executor. This provides a way to assign one or more
'machines' (e.g. the processor in a software decoder, an input
connector, etc) to a particular thread, automatically destroying
the thread when all assigned machines are destroyed.

A wrapper is used instead of a raw std::shared_ptr to eliminate the
concern that such a pointer might be null (fmlib::Thread always
contains a valid pointer to a loop/executor struct) and to provide
better ergonomics.

Test: new unit test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598282
Original-Revision: 3239c3596f494a8a7ffd991a68ded68b5eeeb15e
GitOrigin-RevId: 456401955db3c8b442f2fbc71bd9a7fd477e3e13
Change-Id: I802e2278ab07f05f4ad06d7490313989cf0fd29f