[roll] Roll fuchsia [vfs] Vnode::Serve take raw zx::channel
In fuchsia.io v1 the `Open` call takes a server end of
`fuchsia.io/Node`, and as such, Vnode::Serve also accepted a
`fuchsia.io/Node`. However, this was not entirely accurate: it's
possible to serve a custom FIDL protocol, as in the case of
`fs::Service`.

fuchsia.io v2 in fact defines the server end during Open to just
a raw channel. This CL updates the Vnode::Serve to align with that as
well, which simplifies a lot of call sites.

This was mainly a yak shaving when migrating vfs to use typed LLCPP
channels.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597825
Original-Revision: 47f7b625425997b31a7192881a92fe3eaebe8cc7
GitOrigin-RevId: 5571170b5040466bc1c40d0a6a87dc1f1fe474a4
Change-Id: I77deaa4e3ddf58582694a6ea9431792530218a33