[roll] Roll fuchsia [debugger] Mark attached processes/jobs in "ps"
It can be helpful to see which processes you are currently attached to
when running "ps". This annotates attached jobs and processes in the
output with a margin marker and uses bold so that they are easier to
identify.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597382
Original-Revision: d3558adf0fca9020bee0ce6ccbd3ba09ad727d9b
GitOrigin-RevId: 77d4b8b9fe7dc779d380a6c26b8769099d123065
Change-Id: I71b638cf267550b4c03c3f2d58148debb56207a2