[roll] Roll fuchsia [ffx] Add tests for portmap to_string()
fxrev.dev/596703 added a portmap structure to the device specification,
but didn't add tests for the functionality around that structure. A test
for the to_string() function was specifically requested, as it's a
fairly complex transformation. This CL adds that test.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598063
Original-Revision: 39c908497cd7fa5a4255e1ab0ad90410a55ae443
GitOrigin-RevId: 6de84360f65ae7c0b71e4d7db7a885e833ea8a1a
Change-Id: I352527db8d61a66c21eab7d8826c88c05849b0e2