[roll] Roll fuchsia [modular][flake] Deflake session/agent restart tests
The is_running check is not consistent, likely because the same
FakeSessionShell instance is used when the component instance is
torn down and recreated.

Track the total number of times the shell component starts up
so the test does not rely on the timing of is_running being set.

Original-Fixed: 87344
Original-Fixed: 87293
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598341
Original-Revision: bd08665f4e7375ff93d538e081198103a5b037ff
GitOrigin-RevId: 17148bb8396392b6983b05abf156cc03eeafc95e
Change-Id: I0e7978f321d9c274adcb00c06f98c095722cc941