[roll] Roll fuchsia [ffx] Add more info to daemon link failures
There was a failure in the e2e tests that pointed to a failure with the
ascendd link with what look likes a missing file for the socket. This
adds some extra info to the error messages so that if this happens again
it will be easier to figure out what is going on.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/588330
Original-Revision: 8982a571f96c679c8df31fc747baacbf010601d4
GitOrigin-RevId: 9e3a9c18a333995472b7db664276bd3b792cb04a
Change-Id: I8b393641f747f0b4aa287af347d02870e33002a7