[roll] Roll fuchsia [password_authenticator] Abstract disk ops behind DiskManager trait
Instead of AccountManager being parameterized by a `PartitionManager`
trait, the entire disk management layer is represented by a
`DiskManager` trait.

This allows for the mocks in AccountManager to be substantially less
opinionated about the actual bits on the disk.

Test coverage of `DevDiskManager` is more complete because the mocking
is done at the FIDL layer, which guarantees that the client side of
`DevDiskManager` is fully tested.

Original-Bug: 86859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598302
Original-Revision: 2c491ecd21d639e8c68e4131f97f7272c1f2e59b
GitOrigin-RevId: 5eeef9c8b7ba596c7ba3b5b98c508c7148e8998f
Change-Id: Ic70ec0ee95327b25c4d4860271f72dde94218284