[roll] Roll fuchsia [kernel][lockup] Dump context for all x86 CPUs
On x86, when a CPU is unresponsive, dump the context of all active
CPUs in addition to the one that's unresponsive.

Use Backtrace::PrintWithoutVersion to eliminate redundant version
information (and maps) to reduce size of output.  The version
information is already printed by the panic.

Original-Bug: 86717
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597826
Original-Revision: ba57419fcb61595b3974ecab6fbbdc34a9b9a8cc
GitOrigin-RevId: 300be2f9b94fdb842fed416ef7c41e1f04f5d432
Change-Id: I770f537a3000399c6ff7d351396f8734de9b235a