[roll] Roll fuchsia [ffx] Add example device_spec manifest file
This is an example device_spec, a JSON file which contains emulator
configuration values. The spec is still in active development, so it is
likely to change over time. We plan to follow up with a test which
consumes this test file and generates command lines according to its
contents, but this will be in a later CL.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598741
Original-Revision: fe7c1e284a6cdd63b8d5cee5442761dc4f7dd4cd
GitOrigin-RevId: 49042a49937f14b36aad7f64c5f207c8b709e588
Change-Id: I30dc0560712af551688edc21a719de30f44b9351