[roll] Roll fuchsia [core-tests] Fix flaky test for CPU perf scale bandwidth.
Change the SystemCpu.ScaleBandwidth test to ignore runtime in excess of
the expected value, which may happen in emulated environments on shared
host infra. This is valid because the test is meant to detect receiving
less than the expected runtime when the CPU performance scale is < 1.0.

Original-Fixed: 85846
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597052
Original-Revision: bdc3ce8cdb856e3460db74e77c2bd73b59436675
GitOrigin-RevId: e03a72f28fb8d92764add91cc7a373a94a350856
Change-Id: Ic318899e63218a3c43276305d91041f96f437fb4