[roll] Roll fuchsia [driver_runtime] Add per-thread driver call stack tracking.
This will be used to implement re-entrancy checking in the
dispatcher. We need to know which driver owns a dispatcher,
so every time we enter a driver we will push it onto the
call stack, and check the top of the stack when
fdf_dispatcher_create() is called.

There is a new internal.h API included in the runtime shared
library that is restricted to the driver host. This is so it
can provide the runtime driver updates on driver entries, as
driver host threads are separate to runtime threads which
the runtime has no access to. e,g. e.g. the client could create
a fdf_dispatcher in a bind call.

Original-Bug: 85577

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596062
Original-Revision: 395d9104a3664c062cb67f0b70e1823c36a78c70
GitOrigin-RevId: 8b589889df660d3efe5133cf622d8fe8ca07d725
Change-Id: I972898462271b6a9f1f0e29107bd9ae4fe76a07b