[roll] Roll fuchsia [ffx] Make trace list-categories better
There is a FIDL API called GetKnownCategories which returns a list of
known tracing categories. This is a hard coded list which is in a
config.json file in the trace manager code. It is a subset of categories
which may be available for tracing.

Separately, ffx trace by default enables a list of categories which
intersects the list described above but neither is a subset of the
other. Therefore what should `ffx trace list-categories` show?

This CL adds the default categories to the output. In the human readable
format, I have `Known Categories:` followed by the list returned from
the FIDL call, then `Default Categories:` followed by the constant list
that is defined in the plugin. In the machine output we add known and
default bools to describe where that category comes from. This lets you
have maximal flexibility as a consumer and you can decide what you want
to do with that information.

Also there were tests in the args.rs but they were never run because it
wasn't specified in the BUILD file. This makes those run and adds some
more.

Original-Bug: 86495
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595146
Original-Revision: 77f2f11c91214e9387fac669dcb3d8a152a65742
GitOrigin-RevId: 2f48e071365ec8298d88da45c91b7a21e8510046
Change-Id: Ia45485f203b88fd68980f8362f19bbf53e8bfae6