[roll] Roll fuchsia [wlanif] Only set AP online after start confirmed
An AP interface had its ethernet status set to online as soon as the
request to start the AP came in. Additionally it would also remain
online even if the start confirmation indicated that the start failed.

This change makes it so that an AP interface is only considered online
once a successful start confirmation is received. On a failed start
confirmation the interface is set to offline.

Test: fx test wlanif_unittest
Original-Bug: 86875
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/595447
Original-Revision: 316961ae86a42594e82e6606f046059595f3b770
GitOrigin-RevId: 73642f88f63f9ce5bafe34129c5650ad784d44f4
Change-Id: I597d1109107207bb5ac49b8eb3904e22d6663ce8