[roll] Roll fuchsia [password_authenticator] Provision a new enrypted account
Implements account provisioning by finding the account data partition
and binding the zxcrypt driver to that block device.

zxcrypt is then used to format the block with a key of zeroes in the
keyslot 0.

The last step in provisioning is to format minfs on the encrypted block
device, which is not implemented in this change.

Original-Bug: 86859
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/597832
Original-Revision: 5d77451aff6945e5edb99a5e5992705084b63f08
GitOrigin-RevId: ba1e10123ec255e9e076452966120a7225a1e76a
Change-Id: I0ae8fc9ed65e4ce4f9035a09bd4ea0923936beb0