[roll] Roll fuchsia [wlanif] Support notifying driver of online status
Some drivers, especially those with separate data planes, might need to
know about when the online status of an interface changes. Currently the
wlanif driver handles this and the underlying driver does not know
anything about this.

Add a method to the FIDL interface to allow wlanif to notify the driver
if it's interested.

Test: fx test wlanif_unittests
      build bringup.x64 with drivers-only
      build bringup.arm64 with drivers-only
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/593262
Original-Revision: 1b4c09035c6304184ae84772bc819f59c1a16b49
GitOrigin-RevId: 96c2fa53447fb03876924981906cee905a6e90e0
Change-Id: I1a816ea82da92d20c104a4fc5723f283c01f0dbb