[roll] Roll experiences [workstation_session] Define a storage capability.
This defines a storage capability that workstation session may offer to
components in the session. The storage capability is backed by a directory
in the outgoing path that is populated when workstation_session
successfully retreives a directory from the account system.

Testing: The backend work for this is not complete yet, but I did test
  a variant of this CL that late bound the outgoing directory using
  an isolated persistent storage input. The storage example component
  was able to successfully read and write using the storage capability.

Original-Bug: 81601
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/598654
Original-Revision: 38c20dc1a4770a695132a83c3876c92ddc187e61
GitOrigin-RevId: 967a523cace277a4275fadd77b2b5cc7b4bbc40c
Change-Id: Ia9ce8e9813f7447647d75074093a16e4ccd80171