[roll] Roll fuchsia [ffx] move daemonize tests to a standalone crate
We started to observe a deadlock from this test on Linux, which is
likely due to an improperly handled failure to spawn the child process,
resultant from CLOEXEC behavior (exact root cause was not identified).

Moving this test to it's own crate ensures that the daemonize call
occurs at an appropriately early and isolated place in a program
runtime, comparable with the intended use case of the daemonize()
function.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599525
Original-Revision: 24cca7a4658ed8361c5d0560154aa33887d9928d
GitOrigin-RevId: 9d62cb9b0fb9fedcf767d1762cc1def7514d45b3
Change-Id: I629f38f0fb3e4a9565761b622a0187583dbacc50