[roll] Roll infra/recipes [cl_util] Raise purple if expected tryjobs are not found
In some cases, the tryjobs we expect to query may be missing, either
due to a failure in the scheduling step (possibly a downstream issue)
or an invalid configuration due to the expected tryjobs being
undefined.

This should be considered a failure rather than a silent pass.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/599647
Original-Revision: cb50f4ff3a19a8170df6efc30c69af49fa5b6bea
GitOrigin-RevId: 3bd46006e2304670ab80d032d3108d60b2f6b05b
Change-Id: I9c5a04c93fd174c84b149afbc788ff00756c152a