[roll] Roll fuchsia [wlan][policy] Detect multiple controllers.
The existing way to check if another service holds a policy
controller is to check if the channel has been closed.  This
works on Fuchsia, but is not an option on host tools that
make use of overnet.  This change introduces a wrapper to
test the results of policy commands to check for the epitaph
that the policy layer sends.

The following tests were run to ensure that the error was is
printed when another service holds the policy controller:

Test: donut client scan
Test: donut ap stop --ssid asdf
Test: fx test donut-tests
Multiply: donut-tests
Original-Bug: 87482
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598403
Original-Revision: 1c760b4c78677984dd071fe6e3164df2a4a504a1
GitOrigin-RevId: becf9905d4236b16cc4351d806b2d34bc1364157
Change-Id: I50b186e8d38551a8128c2667610571a9a5eac2d4