[roll] Roll fuchsia [iomisc] Make ReadUntilMatch easier to call
Have it take a reader and targets to match and construct a
MatchingReader itself, instead of requiring the caller to construct and
pass in a MatchingReader.

Also add a new ReadUntilMatchString wrapper function for the common case
where we want to match a string, and it's annoying for the caller to
need to convert the string to a byte slice.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598701
Original-Revision: 8dd986a0ded051260a6b927915ee98a3b4ea29ae
GitOrigin-RevId: 9ebf241e731886ccce2dcc32b1b62a690d2da1fb
Change-Id: Ibcc62569d01b6546312102e1d8449cd2bd591691