[roll] Roll fuchsia [intel][wlan] Make all devices have the same parent
Currently the iwlwifi creates a root device
that is bindable then adds children to that
device.  This is a behavior that is incompatible
with the next iteration of the driver framework.

As a workaround to avoid this behavior, this CL
causes the iwlwifi-wlanphyimpl to add children to
its parent instead of itself.

Original-Bug: 84362
Test: device enumeration test, manual testing.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596702
Original-Revision: 248813f1ec4defe02e31d8e79c4d837c3e57a176
GitOrigin-RevId: ec714d0be2187af001d347490df6a3d32d0d1e67
Change-Id: I9b7afb5a6e07f82ec3a099e0db5f0570fc210255