[roll] Roll fuchsia [fidl][cpp] Add new WireSyncClient APIs and begin soft transition
This CL adds a "->" operator to WireSyncClient which returns an
interface for making managed FIDL calls, and a ".buffer()" accessor
which returns an interface for making caller-allocating calls.

It also starts moving WireSyncClient to be much closer to WireClient:

- Deprecates accessors that expose a raw pointer/reference to the
channel, with plan to fully remove.
- When binding a client endpoint, the client endpoint must be valid.

These in turn open the door to growing WireSyncClient to be more
featureful than a simple ClientEnd wrapper in the future (e.g. unbinding
and closing the endpoint on error, terminal events support, ...).

These changes are started in one CL so that we don't have to do two
separate waves of user code migration afterwards.

The "->" returns a SyncEndpointManagedVeneer, a newly introduced veneer
object that only exposes managed FIDL calls. This is so that the user
doesn't accidentally write the following:

  fidl::WireSyncClient<P> client;
  // The first arrow should be a dot.
  client->buffer(arena)->Foo();

An earlier idea was to separately track two pieces of state in
WireSyncClient:

- Whether the client was default constructed/initialized with an
endpoint.
- Whether the endpoint in the client was a valid handle.

While theoretically cleaner, in practice it caused more confusion as
it's very tempting in C++ to treat a default constructed channel as an
absent channel. Consider the following snippet:

  zx::channel client_end;
  if (some_feature) {
    client_end = GetClientEnd(...);
  }
  fidl::WireSyncClient<P> client(std::move(client_end));

Here the |client| might be initialized with an endpoint, but the
endpoint was invalid. The user's expectation was that the client should
be absent when the endpoint is absent, hence we preserve that
expectation in this CL.

This CL migrates the FIDL test code to the new API, along with the bare
minimum of user code changes that is required to compile and pass tests.

Original-Bug: 85688

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/598333
Original-Revision: 935909641ea0b12b6973acb9e9b5bef2088d54ca
GitOrigin-RevId: ca731184cb29382da931a3c183488cccebc56959
Change-Id: Id92976e36a8b534795c53a63928d75047e0022ae