[roll] Roll fuchsia [ftl] Add diagnosis output to volume extractor.
Adds a diagnosis section for analyzing known issues in an FTL. This
makes it easy to reach inside the FTL and check on various conditions to
identify known symptoms of issues that have been found to corrupt an FTL
volume. Including a check for Partial Page Writes, as well as
specifically checking for them when the fix has already been applied.
Also looks for vpage mapping collisions to identify Premature Block
Recycle.

A couple of problems have just had their logging adjusted to make
spotting the issue more clear, such as the failed read during load time,
and also a vpage mapping that points to a non volume block which would
indicate Premature Block Recycle.

Also cleans up some of the logging output, like iterating the correct
number of vpages when dumping to remove the harmless error log, and no
longer logging when bad blocks are found, as that is a normal and safe
occurrence.

Tested: Against 4 different images with known problems in them, and 1
image that was doctored to induce PPW with the fix.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599265
Original-Revision: 402af9849902e259de6ee81be76cf6b548069f44
GitOrigin-RevId: 5661d462f3e857a6ff667073a3ab62308896cc71
Change-Id: Icdb56b3cc981e0968ff1405dce37217f09f9f56a