[roll] Roll fuchsia [kernel][vm][ppb] StackOwnedLoanedPagesInterval
This RAII class will allow a thread to wait on a loaned
page to no longer be owned by another thread's stack.

The OwnedWaitQueue is used for proper priority inheritance
so that the waiting thraed doesn't get stuck waiting for
a lower-priority thread.

This CL stops short of making the PageQueues changes needed
to keep a backlink set while a loaned page is in a VmCowPages.
That'll be in a later CL.  For now, there won't be any loaned
pages so the clearing of the backlink sometimes while a page is
still in the same VmCowPages is still ok for now.

Tested: vm-stress-test, CQ, vmo_stack_owned_loaned_pages_interval_test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/591623
Original-Revision: a589ed304c2fd8c61c72d959df62276372c7f7c2
GitOrigin-RevId: 2163395d54087ac412e11345d388dc51f8b6fb92
Change-Id: Ida229297e932caadd99a92a6bf54ba76aaf2c664