[roll] Roll fuchsia [usb][xhci] correct segment_index_ logic
segment_index_ is written to the DESI bit of ERDP and should be the last
3 bits of the offset of the current ERST entry (see 5.5.2.3.3 in XHCI
spec). It is used to quickly determine if the Event Ring is full.
Originally, we were increasing segment_index_ only when we wrapped
around from the last buffer to the first buffer, but we should be
incrementing segment_index_ every time we move to the next buffer (i.e.
segment).

Test: tested on NUC (17 hours and mouse is still working)
      tested on Atlas (mouse isn't laggy anymore)

Original-Bug:81262
Original-Bug:78895

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599861
Original-Revision: 01187fe097bf5d0bb0a4a8f39de82041e3a84475
GitOrigin-RevId: 17a9cdc9a8799d36a65212394e8bd86fbfd858ba
Change-Id: I8bb68820a06fa17969362f5d01d7052bdb85a527