[roll] Roll fuchsia [build] Add vpython3 to tool_paths.json
So the sdk recipe can use vpython for invoking SDK scripts. I was hoping
to be able to stop using vpython entirely, but that causes some
breakages on Mac, so instead I'm going to just make sure that the recipe
uses the versions of vpython3 and python3 that are included in the
checkout.

Original-Bug: 77963
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601202
Original-Revision: 5986ba07321bca12b320d5852b4ae1f0e5199458
GitOrigin-RevId: dde201cf94587128cfc857fa1328f13854865532
Change-Id: I305fd88851ce82fb0ccca4bdc35304b5bd303a6c