[roll] Roll fuchsia [timekeeper_lib] Replace zx::time_utc with a library-defined type.
Zircon previously exposed a UTC clock but this was removed with the
zx_clock_get syscall June 2021. We now need to clean up to remaining
references to UTC in zircon including the ZX_CLOCK_UTC constant and the
zx::time_utc type.

The zx::basic_time<> type will still exist and so can easily be used to
define a replacement type. This CL defines a replacement in
lib/timekeeper and uses this new type for in-tree clients. The new type
is API compatible, allowing a soft transition for out-of-tree usage of
lib/timekeeper.

Original-Bug: 86976
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599944
Original-Revision: 2259307cbe81cb6eb5694fd57e6f85fa913fbeb3
GitOrigin-RevId: 8ca2ae58d4e5f5b65226270e26ef185c3e4f8e11
Change-Id: I8458787aad93109676e0d8a356aedecaf6d65b32