[roll] Roll fuchsia [ffx][emu]Simplifying FfxConfigWrapper and test usage
The previous implementation used an async trait, and 2 implementations
of the trait, one for tests which was backed by a hashmap, and one for
non-test usage which called ffx_config methods directly.

This change merges the two and removes the async trait. It is much more
simple to understand what is going on, and removes the complexity of
async_traits and sharing them across threads (which I encountered in
a prototype).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/600226
Original-Revision: 30a6266fe8de55c065581ab306c5bee76d95116f
GitOrigin-RevId: 274e39acac4761bea61b38c27116c83f55fa2b55
Change-Id: Iad7807d959eba34704427691b21219c27f2d8cae