[roll] Roll fuchsia [bt][a2dp] Don't start player until data arrives
In some cases when a source stream is connected, it is suspended without
sending any data.  This causes us to start up a Player and immediately
tear it down, which is undesirable.

Instead start the player once the first data arrives in the data stream.

Add an estimate of the total streaming seconds to the inspect data for
a DataStreamInspect.

Test: tests updated
Test: fx test bt-a2dp-tests bt-a2dp-encoding-tests
Test: fx test bluetooth-crate-tests bt-rfcomm-tests
Original-Bug: b/203686738

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599361
Original-Revision: d8c0a274630e653d2d42cfb05810d035ad616d12
GitOrigin-RevId: c3360013f6ff3bc8ef0954240e19429ad519d75f
Change-Id: Id37441dfbd4aa665a728220b3cc708d3729b1a42