[roll] Roll fuchsia [ffx] Allow specifying a session with dump
This also makes the flow for handling offline targets in
StreamDiagnostics clearer: if the request is for `dump` mode (i.e.
StreamMode::SnapshotAll), then we search for offline logs with the
specified session offset. Otherwise, we need the target to be up.

Note that tests are missing from `daemon.rs`. These will be added once
we move this code to a service as the current setup makes this very
tedious to set up tests.

Original-Fixed: 85060
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592155
Original-Revision: 56a1c085605e7459e20d642b82ecc74f0f7c8fbf
GitOrigin-RevId: 1831d14f67b07f97e72e9adf092026474295b94a
Change-Id: I153882cdb59b317b27512dfe6051079984671581