[roll] Roll cobalt-registry [cobalt][storage] Add metrics for bad blocks and nand ECC.
Metrics to examine the failure bahviours of the nand reads with regard
to ECC, to get an idea of how well/badly the fleet is faring. These will
be populated by the Nand driver reporting the results of read attempts
via Sampler.

Metrics for the number of bad blocks in the NDM, this will be reported
by the FTL/NDM process at start up and updated after each request along
with the existing stats to keep it in sync with the current view. This
will also be populated to inspect and reported via Sampler.

Original-Bug: 85321
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/cobalt-registry/+/598026
Privacy-Approval: Griffin Boyce <glamrock@google.com>
Original-Revision: 5d0181eeb6f81802f39146f84958e6dec1b8c37e
GitOrigin-RevId: 5ba04b5452a7f956af36578f3ebfa0bdf961a914
Change-Id: I2289cfac421d31e3f6480e057e70612f3985f7f5