[roll] Roll fuchsia [rust_toolchain] Remove `must_not_suspend` lints
The recently-added lints have been made unstable due to some false
positives. Let's remove them for now and replace them with a roughly
equivalent comment. Once the lints are re-stabilized we can remove these
and replace them with allows instead.

These lints will block the build if they're not allowed on the current
toolchain, but they'll also block the build if they are allowed on the
next roll. Because of these conflicting requirements, this change also
adds a temporary escape hatch to get out of the migration trap. This
will be removed after rolling the toolchain.

Test: Manual, ran on local workstation.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/600841
Original-Revision: a815a04ec0afb54012d08c8bdde56eabf529b062
GitOrigin-RevId: aa80a098554557621953e2f9e27c472f7c0f55fb
Change-Id: I5ed54d868bf99ba82f8e145dc9cb99f7a6f74d0c