[roll] Roll fuchsia [bt][rfcomm-integration-tests] Fix flake in test
The RFCOMM integration tests occasionally flake when validating that
data can be written to one end of a RFCOMM channel and read from the
other end.

Fix this flake by reading in the receiving end until data is produced
(or a test-defined timeout is hit). This matches the behavior of the
bt-hfp-audio-gateway-integration-tests which experienced similar issues.

Original-Bug: 82484
Test: fx test bt-rfcomm-integration-tests
Multiply: bt-rfcomm-integration-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599271
Original-Revision: 30ca7b0591c0cc7605ffd529c392b40ec7bf5924
GitOrigin-RevId: 78abba6bdc08eb8b22fe21fb00f3b33697eab259
Change-Id: Ie1ca32e070fece95b88c3f2db1a2fc8ed99a372a