[roll] Roll fuchsia [loader_service] Migrate to new WireSyncClient API
This is part of a migration to stabilize the LLCPP API. In this step, we
are aligning the interface of WireSyncClient and WireClient. The main
changes to WireSyncClient are:

- Similar to WireClient, use the "->" operator to make FIDL calls with
managed memory allocation.

- Similar to an earlier change in `fidl::WireCall(...)`, use the
`client.buffer(arena_or_buffer_span)` accessor to make FIDL calls with
caller-controlled allocation.

- Removing accessors that expose a raw pointer/reference to the
channel, instead providing a `Bind` function to initialize a default
constructed client, and a `TakeClientEnd` function to de-initialize it.

- When binding a client endpoint, the client endpoint must be valid.

See https://groups.google.com/a/fuchsia.dev/g/fidl-dev/c/nF72xpPL4gU

Original-Bug: 87683
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/600540
Original-Revision: 7abc757ecc5a112ee030611590225e7986867639
GitOrigin-RevId: 5199b0d00c521f827264210be6e570605f04ab25
Change-Id: I311f6b791a88fb3702c348963022ee9dc7053e66