[roll] Roll fuchsia [assembly] Add signed support for images manifest
The zbi in the images manifest may be signed, and we need to pass this
bit of info to the create-update operation, because the entry in the
UpdatePackage depends on it (zbi vs. zbi.signed).

Also add the signed zbi to the images manifest, instead of the unsigned
one.

Original-Bug: 84904
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601011
Original-Revision: 964d46b159450919c6e914d880231dde184e5985
GitOrigin-RevId: 9168671d4801d6ced1cd2dfc410d629de6421a6d
Change-Id: Ic8030c21c7daba401ef9ef538011cc9591ede5fe