[roll] Roll fuchsia [test_manager] Wait for client to close suite controller
Holds the SuiteController channel open instead of closing it after
all events are drained. This is a workaround for an issue where overnet
returns ZX_PEER_CLOSED if the server returns an event, then immediately
terminates the channel. See fxbug.dev/87890.

Note that the SuiteController API explicitly states that the server
will drop the channel after events are drained. This breaks that API,
however, none of the clients currently rely on this behavior. We'll
need to remove this workaround once the bug in overnet is corrected.

Multiply: test_manager_test

Original-Bug: 87976, 87890

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601524
Original-Revision: 6746572ce5bf6474bee6937ce03a6bb39e81d3a4
GitOrigin-RevId: 5668ce413c16b0b5e3766db5a8197ba3a0e277b9
Change-Id: I64838659edb9fe3e66cf0d79da2070afa37e61fd