[roll] Roll fuchsia [rust_toolchain] Remove hack to escape migration trap
Fixes: 87730

The details of how to perform this hack again could be left in, but
it's probably for the best that we prevent this class of migration
trap from happening instead. They'll always still be here in git
history.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602561
Original-Revision: dac2835c5409b44c56fe977c1d4da06bc2e7cf77
GitOrigin-RevId: 46eaeb7c24a90958dc10d22507ca7e234d0294d2
Change-Id: I2903eee27344db2cbe8756ab5ddb83c5d6e65040