[roll] Roll fuchsia [zircon][vm] Get backlink until after pages removed.
To chase down a page that needs to be reclaimed, we need
to be able to get a RefPtr<> tally on the owning
VmCowPages until after the page has been removed from the
owning VmCowPages.

This avoids the page reclamation loop spinning between
VmCowPages refcount 1 to 0 and ~VmCowPages removing the
pages.

To do this, we separate the VmCowPages and its container
to allow each to have its own refcount.  The VmCowPages
refcount can drop to zero while the container's refcount
is still 1.  During this time, we can still get a ref to
the container, and via that ref we'll be able to replace
the page.  Before VmCowPages removes its 1 ref tally on
its container, it has fully removed all pages to FREE
state.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601301
Original-Revision: 4479602142ef7eb5783b8161badeb3b2e1cb2feb
GitOrigin-RevId: edbfa1d3d3c31eb66ba2346016815ac0ec90e43b
Change-Id: I6778d0d3ed9fd27fed929147467e416b8fafd56a