[roll] Roll fuchsia [ffx] Reorganize the emu plugin to support engines
We've had a lot of conversations about how the emultator plugin should
be organized lately, as we prepare to add multiple instances and
multiple backing engines. We settled on an object-oriented design where
each engine is its own type that implements a common interface, and the
engine used for a given instance comes from the manifest type selected
during launch.

This CL introduces a new layout for the code in the emulator plugin
source tree which reflects the engine-based design. The existing code
becomes the "vdl" engine, which uses device_launcher as the backend.
Common code, such as the port_picker library, remain in the common
library.

A few skeleton types are also added, which will be populated and
plumbed through in the next CL, for the common config types that will be
sent to each engine.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602565
Original-Revision: e0084d1633d4bc1ff59411c079110adc8aa7d59f
GitOrigin-RevId: 3371ee922ae10534709391e2718974caec8daff4
Change-Id: Ie41b5bff3ae7a7ef25b3c958b2082aa5e5559fb5