[roll] Roll fuchsia Revert "[fidlc] Record dependency edges from attribute arguments"
This reverts commit 97ae98ffc2e10c13bc588463713c72d1cb623cb1.

Reason for revert: This was actually a bad idea, because it leads to
an unresolved constant error for HEAD in the @available attribute.
When investigating why this doesn't happen for MAX bounds (which is
implemented with an identical hack), I realized we don't record
dependency edges for constraints. So there is no need to need to do
so for attribute arguments either. Backends like fidlgen_hlcpp
already emit constants before anything that could use a constant.

Original change's description:
> [fidlc] Record dependency edges from attribute arguments
>
> This CL makes the topological sorting step in fidlc take attribute
> arguments into account. Omitting them is not necessarily a problem
> for any backends today, but cycles will cause compilation errors
> anyway, so we might as well record the dependencies.
>
> Test: fx test fidl-compiler -- -f "DeclarationOrderTest.GoodAttributeArgConstantComesFirst"
> Test: confirmed new test failed before flat_ast.cc change
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599581

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602805
Original-Revision: 5fd72f169813ca7d7300275fe02e0b5055d40a20
GitOrigin-RevId: 621faf97ab3fbead830278828a4a4a01a1c88a7f
Change-Id: I7a4542c709ceca4e5aa650ec920004879c856d83