[roll] Roll fuchsia [flatland] Defer response to ChildViewWatcher.GetViewRef().
The previous behavior was to respond to GetViewRef() as soon as the
ObjectLinker connected the view/viewport.  This resulted in the
possibility of the parent immediately attempting to set focus on the
child view, before the child appeared in the ViewTree, and therefore
focus would not be set.

Original-Fixed: 81850
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601921
Original-Revision: 01ee0e421d61d3a64fc169a92ac25e257d3e40a6
GitOrigin-RevId: ae0821435ca1ca0fcb907c62f99cfaf649ee8963
Change-Id: Ica048c26a337abdefb15ed9dbd3331248ede320e