[roll] Roll fuchsia [starnix][msync] Add validation to unimplemented msync
The gvisor mremap tests use msync as a way to check whether a page is
mapped. They expect an ENOMEM error when msync is called on an address
range that has one or more unmapped pages.

This change implements this validation, but still logs that the syscall
is unimplemented because it doesn't actually do anything besides the
validation.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602811
Original-Revision: 78c9c1deba9f5c116d9122840dd3f3b340327f02
GitOrigin-RevId: a27e44ffbd0b55ecf87551cc721c1a10d7f05ca6
Change-Id: I6010c78ccab5c5649c37d56335f3f5ab31382dc7