[roll] Roll infra/recipes Reland "[sdk] Use prebuilt python interpreter from checkout"
This reverts commit 45c6d347f9dc1bdbfce996947264e891fed0e6ac.

The original CL switched the recipe to use the Python 3 interpreter from
the checkout instead of vpython. This broke the Bazel SDK tests on
Mac, causing the Bazel build to fail with a permission error from within
vpython when it tries to create a virtualenv. This was pretty
mysterious, because we shouldn't have been using vpython at all anymore.

My hypothesis is that the breakage happened because:
1. Buildbucket builds include a `vpython-native` CIPD package that
   makes `python` an alias for `vpython`.
2. The Bazel SDK tests have `python` shebangs, so they end up using
   `vpython` when run from within a Buildbucket build.
3. When we were running the top-level command using vpython, vpython
   would immediately create the virtualenv so it wouldn't need to be
   recreated by subprocesses. So when we stopped using vpython directly
   the virtualenv would not exist when vpython was run by Bazel, so it
   would try to create it and fail. (I'm admittedly not sure why it
   failed there but not in the top-level invocation.)

There's nothing Mac-specific about this situation, so I think the same
thing was happening on Linux and we just didn't notice it because the
the virtualenv creation only failed on Mac for some reason.

The solution is to add the prebuilt `python` and `python3` executables
from the checkout to the front of $PATH prevent the SDK tests from using
vpython instead. This is only strictly necessary for the `run bazel
tests` step but it can't hurt to also do it for the other steps that run
SDK Python scripts.

Original change's description:
> Revert "[sdk] Use prebuilt python interpreter from checkout"
>
> This reverts commit 7f89b7e3df96045284529c2d7f3ef754101058ad.
>
> Reason for revert: Fails on Mac
> Original-Bug: 87751
>
> Original change's description:
> > [sdk] Use prebuilt python interpreter from checkout
> >
> > Previously we ran SDK scripts using the version of vpython2 that was
> > installed ambiently on the bot. Using the prebuilt Python interpreter
> > from the checkout will make the SDK scripts slightly more hermetic and
> > also let them drop support for Python 2.
> >
> > Depends on fxrev.dev/599124, which adds python3 to tool_paths.json, and
> > fxrev.dev/599341, which updates the SDK scripts to support Python 3.
> >
> > Original-Bug: 77963
> > Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/599142
>
> Original-Bug: 77963
> No-Presubmit: true
> No-Tree-Checks: true
> No-Try: true
> Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/600650

Original-Bug: 87751
Original-Bug: 77963
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/601321
Original-Revision: bc0cda67f5969e60ce72e2510777c594b4d3a72e
GitOrigin-RevId: 8545750936d1900fc06835fce55d1b4acef63f64
Change-Id: Id21266ed4fe71a317a7f17d695c3d71a344763ec