[roll] Roll fuchsia [fidlfmt] Adjust comment indentation algorithm
The algorithm intends that a standalone block comment must be
indented to match the indentation of either it's immediately preceding
or immediately following line, whichever is greater. The previous iteration
of this code, however, was only indenting the comment to match the
indentation of the following line, completely ignoring the preceding
one.

Additionally, `find . -iname '*.fidl' -not -path "./out/*" -not -path
"./prebuilt/*" | xargs -I {} fx format-code --files={}` was run to
ensure that all in-tree files incorrectly formatted due to this bug are
now correct.

Test: fidldev test fidlc
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602262
Original-Revision: 964a1196ff25317b96172952c941ba95f9520bef
GitOrigin-RevId: f92eff594b6461efce76bc72b50b485dea76485d
Change-Id: Ib2cfcf7cec4a49e942e6a52298306ba52df72d2c