[roll] Roll fuchsia [fidl] Fix typos and markup for block device FIDL.
Declare `BLOCK_VMOID_INVALID` to be of type `Vmoid` (= `uint16`), not
`uint32`.

Set type of fields in `block_stats_t` from `size_t` to `uint64_t` to
match the FIDL. `size_t` is technically not guaranteed to be `uint64_t`,
although as a practical matter it is on all Fuchsia systems.

Update the comment explaining the `flags` field of `alloc_req_t`. It
took some spelunking to find the actual live definition. It is still not
super clear; perhaps the flag name(s) should be in an `enum` with a
suitable name, or prefixed with `ALLOC_REQ_T_FLAG_` or similar.

Original-Bug: None
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602109
Original-Revision: d203feda7ea2c30199c2a8eba746f962c4d0ff99
GitOrigin-RevId: 78083f8c599df748784275d1c872b0ccb7bd806c
Change-Id: I681e132a064b2e5e0ab8b8f9b61757a25dcec448