[roll] Roll fuchsia [ffx] Remove package serving from ffx emu
Previously, ffx emu was built to start the package server for the user
alongside the emulator. As we redesign the launcher, we want users to be
responsible for starting it themselves. In August, we made running
without the package server the default and left a note to completely
remove it after a month. It's now been 2.5 months, so the time has come.

This CL removes the package serving flags from ffx emu, making it
impossible for a user to start the server from here. They must use
something like `fx serve` or the upcoming `ffx repository` instead. This
will not impact infra tools that are already starting it themselves or
using other launchers.

Original-Bug: 82804
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601541
Original-Revision: f512418efc20a2e667d0da5e8c3e778524395cb6
GitOrigin-RevId: 84a2ae2ea667aa3ebe83f81db24167507652e2e8
Change-Id: I91be7a50529d11b231e3652b65b5ebb16e606634