[roll] Roll fuchsia [remote-control] Close tunnel if one side closes
While previously we would shut down the tunnel if the zx socket was closed,
we wouldn't shut down the zx-to-tcp task if the tcp socket was closed. This
could waste resources, as it'd leave the zx socket up and running, and thus
the tunnel to the host.

This fixes that by using emitting a () from the tcp-to-zx task over a
channel to the zx-to-tcp task, and adds some tests to verify this logic.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/603274
Original-Revision: 6846010df9a56aa2aa25054b4ec73c7a9a84e75b
GitOrigin-RevId: 96d4887473149cef5e66c6aeb24eb70899440cd7
Change-Id: I69952b502e4626e55a0f03cf7a407bedae8de992