[roll] Roll fuchsia [carnelian] support flexible facets
In order to support use cases like the background of a button or menu,
allow facets to have flexible sizes.

It turns out most of the support was already in place. The main change
in the layout portion of this change is to pass the size of a facet as
calculated by layout to the facet during render.

Previously the entire view size was being passed to the Facet trait's
update_layers method. Now it is being passed the size layout would like
the facet to be. Facets that are flexible about their size can use that
size as their target.

The only facet using the size parameter was the textgrid facet in
virtcon, which I switched to using the view size from the view context.

With this change it was useful to implement the member data for the
Stack layout, which lets one set the top, left, bottom, right, width
and height of a Stack group member.

In order to implement Stack's positioning features it was necessary to
pass group member data to the Arranger's arrange method as well as the
calculate_size method that was getting it before.

Testing: existing layout unit tests, plus some new features in layout
gallery for some eye tests.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/599941
Original-Revision: f3519745c833309426673de618e64e6b7c30f2c6
GitOrigin-RevId: 5e3ae7fa517db705b16b21f6869499593de32720
Change-Id: I44dcde238defb26aa67f1fbb4516a8c8f0a844cf