[roll] Roll fuchsia [cflags] Make FIT_DECLARE_THREAD_CHECKER -Wextra-semi compliant
This macro emits either a member declaration or nothing
depending on NDEBUG.

The declaration needs to terminate with a semicolon,
whereas an empty definition cannot have a dangling
semicolon after it or it triggers -Wextra-semi.

This leaves two options:
1. Change all usages of FIT_DECLARE_THREAD_CHECKER(...)
to not terminate in a semicolon.
2. Surround the no-op definition of FIT_DECLARE_THREAD_CHECKER
with a suppression for -Wextra-semi.

I chose the second option, for consistency with
FIT_DCHECK_IS_THREAD_VALID which is both terminable
and terminated with semicolons at every usage site.

Unfortunately the same trick used by
FIT_DCHECK_IS_THREAD_VALID isn't valid here,
because ((void)0) only works in a function body
and not in the context of a declaration.
There is no no-op declaration spelling that can be
terminated with a semicolon.

Original-Bug: 56257
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/603581
Original-Revision: f5499c26a237476e281dc94dcfbff265227b4a01
GitOrigin-RevId: f2d8847811dd49982a927d761e9336453c5a00ef
Change-Id: Ia4703bdae00a4eb3fcd26a5644fc1f78cde80f8b