[roll] Roll fuchsia [iwlwifi] Better handling the time event removal
The original code removed the time event during the associaion stage.
This is okay for most cases, but not always correct.

The correct way to handle this is relying on the
TIME_EVENT_NOTIFICATION from the firmware, which tells us when the
time event starts and ends. So, the new code re-enables the driver
code to correctly handle the notification.

Original-Bug: 87392
Tested: Added a test case to cover it.
        On 9260, with v38 and v43 firmware. Both works pretty well.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601531
Original-Revision: 7dba5a712a0c6c125dc2b8f4c62058760aa98f13
GitOrigin-RevId: fae3461f21447c9397cebc594510a42e7f6d3134
Change-Id: I69f3868177aa272bf3983c8572fbe8700af41221