[roll] Roll fuchsia [llcpp] Remove transport_type field from C messages
This CL moves the transport_type field from the C message structs to the
LLCPP IncomingMessage and OutgoingMessage classes. The idea is to
require C messages to always use the FIDL_TRANSPORT_TYPE_CHANNEL type.

Note: this does not change the type of the handle_metadata field to
match the channel transport type. This can be done in the future, but
it requires removing dependencies on the the message class in LLCPP
classes, which is better left to another CL.

Original-Bug: 85734
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602564
Original-Revision: 7f845aa646714fa04b0a8444d90f9080c28ece84
GitOrigin-RevId: 779a815802e6b108cc8907d52e4e8adc33756d22
Change-Id: I0abea7341141abdd600b3683365ffc839bc39280