[roll] Roll fuchsia [PBM][images] Copy sample virtual device manifest
This copies a prototype of the virtual device manifest into the correct
location for it to be processed as part of the product manifest bundle.

Once the manifest is being generated by the build, this change can be
reverted.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/600054
Original-Revision: b4a2edcf2b5f397d674705a2a354aa89cfc771b5
GitOrigin-RevId: b0fb0d720db3c15e354e36bb29324cb004996d9b
Change-Id: I8e86887ac0c14fdca37aed87fe1083d2e86796b2