[roll] Roll fuchsia [netstack] Add bridging integration test
The test creates N+2 netstacks. The 2 netstack that are always present
are named `gateway` (not actually a gateway since it doesn't route
packets to the outside) and `switch` (not actually a switch, but just
a Netstack on which interfaces are bridged together).  `switch` and
`gateway` are connected via a network. In the simplest testcase,
a single other Netstack is created (named `host1`), connected to
`switch` via a network. The interfaces on `switch` are then bridged,
and the test verifies that `host` can acquire addresses and communicate
with `switch` and `gateway`.

There are also test cases to ensure that link/admin status changes on
the bridged ports do not affect communication on the bridge.

Also add a test which ensures that removing a bridge does not
unintentionally enable an attached interface if it is disabled.

Multiply: netstack-bridge-integration-test
Test: fx test netstack-bridge-integration-test
Original-Fixed: 85075
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/585482
Original-Revision: 9c3bae7dd37d42a451cd5f50057c34d4c6049442
GitOrigin-RevId: 650240c2156ac5d961c1c95e8b84dac9703526f4
Change-Id: I9e6e35586a92ce2be69c10a16a69d8307246245e