[roll] Roll fuchsia [llcpp] Use VTable rather than transport type in message classes
VTable is now used to indicate the type of the transport and allow
easier lookup of fields stored on the VTable. The motivation for doing
this is to avoid a dependency in llcpp_transport.cc on the all of the
transports.

Original-Bug: 85734
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602568
Original-Revision: c21fe82201e410a03397ac283788285a10a48bc7
GitOrigin-RevId: 658389f2379a88c035e725627d82efa9acc68b77
Change-Id: I2b3b3624d912845a4a430356d7e3685134e49314