[roll] Roll fuchsia [opencl] Create OpenCL loader FIDL
This uses the same approach as the vulkan loader.

The interface is discoverable. The OpenCL drivers on Fuchsia implement this
interface is they support being loaded via the OpenCL-ICD-Loader.

The OpenCL-ICD-Loader on Fuchsia searches for this interface to find OpenCL drivers.

fixes: 88249
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/603882
Original-Revision: a661129ac5d1dc7ca8cb42b515be9e7713736587
GitOrigin-RevId: e2be12c77733ec83a6cb8b399fb1c5f51f455963
Change-Id: Id192b406aac85ad15e55e856072a61b8672cc823