[roll] Roll fuchsia [llcpp] Use transport vtable implementation for message write, call and read
This CL:
- Calls vtable->write() rather than zx_channel_write_etc for message
write, similar for call and read
- Adds CallOptions, ReadOptions, WriteOptions to respective message and
transport functions. This was needed for read() now that it is
transport agnostic but the same change was applied to Write and Call
so they parallel one another.
- Removes EncodeFlags / DecodeFlags. These might be needed at some point
but there are no short term use cases.

Original-Bug: 85734
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602569
Original-Revision: ffc96168784a70ca3b273412f5009610339f3d10
GitOrigin-RevId: 4e3a6545e4233f28dbee95f18ce9d4034bdfd217
Change-Id: I8345e71c3d2ce20c9462abeafe88126dd21b00b3