[roll] Roll fuchsia [kernel][vm] Commit and MapRange do not dirty
When a pager-backed VMO is committed, its pages are populated, but not
dirtied. When a pager-backed VMO is read-write mapped with MAP_RANGE,
the mapping is populated, but the pages are not dirtied. Both these
cases should not generate ZX_PAGER_VMO_DIRTY requests.

Also fix up mappings for pager-backed VMOs so that they are mapped
initially with the write permission removed, so that future writes can
be trapped and the dirty state updated.

Test: core-pager

Original-Bug: 63989
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602784
Original-Revision: dd89fda748e3df7761ccc83066d86f61d705c664
GitOrigin-RevId: a37142900a0fec70f5494b4029e3ba891dfde340
Change-Id: I33703f3ab1947b8fb0dc7d0d4b1e6c6fc9eb6cf1