[roll] Roll fuchsia [test_diagnostics] Don't panic on channel closed
Due to recent change to pipeline requests, the channel may be closed.
This occurs because the client closes the channel and drops
outstanding requests when it receives the first error. When the server
attempts to respond with a second error, the channel is closed.

Original-Fixed: 88295
Multiply: test_diagnostics_host_unittests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604223
Original-Revision: 7b171fdf9e01eb30701408541504b0abfc6079c2
GitOrigin-RevId: 1375af14e7f04667293bbc9b7f79128377e91f5b
Change-Id: I9b7e67a491e78f53245c007261b776fa47a69557