[roll] Roll fuchsia [fxfs] Make blocksize 64-bit within fxfs.
This removes a lot of casts from the code and alignment means that we
likely don't pay any memory cost for this either.

Fuchsia will still require 32-bit block sizes but the number of safe
casts required to for that interface is much smaller.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604002
Original-Revision: 007a915cfd02335d3b168019eff9ef9bff7fecec
GitOrigin-RevId: 8a5382a3533b0f72be6859d4835a478587fcc4fa
Change-Id: I5d2673f727d8c1f8efdcf55ea5d26005582dcf30