[roll] Roll fuchsia [setui] Swap all stash commits to flushes
This is a temporary and easily cherry-pickable solution to switch all
commits in SetUI to run flush instead. If the device runs out of space
when committing, stash will close the connection, which causes that
particular setting to be stuck until the device reboots.

Using flush will allow SetUI to continue to function and keep the
settings values in-memory, and return to normal functionality once the
device is no longer out of space.

Original-Bug: 88349
Test: fx test setui_service_tests
      manually tested on device by continually cycling flashlight and
      volume slider
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604647
Original-Revision: 86c3a911314489c3749d523cfc3988a85caa6771
GitOrigin-RevId: 35012c47f76242ba0a840f6661acc023e40b5e80
Change-Id: I45734ab43584679678d437182c03031e3acfe31c