[roll] Roll fuchsia [fxfs] Fix some flush issues
Keeping track of the minimum size was proving complicated and it had
bugs. This CL simplifies things by removing the minimum-size tracking
and making sure that we commit shrinking a file before we grow the file
again.

This CL also addresses some issues with reservation tracking by the
cache. Keeping the reservation sizes accurate whilst allowing writes and
flushes at the same time is quite difficult: if a write arrives after a
flush has started, then the amount we want to reserve might depend on
whether the write is in a range that is still to come or in a range that
has already been flushed. To simplify things, this changes flushing so
that it always flushes *all* the outstanding data and then, to make sure
transactions cannot get too big, we make sure to limit the amount of
outstanding data for a particular file.

Original-Fixed: 87196
Multiply: fxfs-truncate-large-tests 100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/601621
Original-Revision: 406269aa740d84fa72d23051beb029fcc9c2d13d
GitOrigin-RevId: 7d3e89b27ee5f171132e64e8678dd79c23101d29
Change-Id: If45e7f6c5823e6ab01458edf890be79e8152d213