[roll] Roll fuchsia [fxfs] Make opening stores thread-safe
A previous CL erroneously removed thread-safety for opening stores which
was believed to not be necessary, but it turns out that it is. The
graveyard needs access to stores for its initial reap process and at the
same time stores can be opened for mount.

Original-Bug: 88401
Multiply: fxfs-corrupt-tests 100
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605101
Original-Revision: af957e4ee771842c2fbf0d6bc27252ca0c488e42
GitOrigin-RevId: 9b276cff6faaa587c31c4dedf3204877f90ad4b8
Change-Id: I2384e8f659ca44d11c3515b06b8393e7868946a6