[roll] Roll infra/recipes [autocorrelator] Skip if there are no gerrit_changes
I believe this is the only place in the code path of the fuchsia.py
recipe were we assume that gerrit_changes is populated for tryjobs,
which makes it impossible to trigger a tryjob manually using `bb add`
without -cl as you get failures such as
ci.chromium.org/b/8830925222530391713.

`bb add` with no -cl argument isn't a workflow we officially support,
but we might as well support it as long as it's easy to do so.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/infra/recipes/+/604263
Original-Revision: 186e9b33eebfc06d74ca04a2abc05a3a9854485a
GitOrigin-RevId: f0ba26e372b4f326db33aec5be4e2e25029b8dc8
Change-Id: Ic7874a5b955e1a7910cc02237e1d8ed85cbf4fc2