[roll] Roll fuchsia [fidlc][c] Emit coding tables for request/response payload types
This CL removes the conditionals in fidlc suppressing coding tables for
request/response payload types.

Given a method `mylib/Foo.Bar`, the C request type for the method is
`mylib_FooBarRequest`. However, this conflicts with the name of the
request/response payload type that is synthesized by fidlc. This CL
renames the message types (header + payload) to RequestMessage, and
similar for responses and events.

Some low-level libraries that are using the C bindings in a bare-metal
way are updated. The coding table is an internal API, hence should be
API compatible to rename.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604069
Original-Revision: f8005dcde2a89556981cdcbb2b20c19d4373747b
GitOrigin-RevId: deaf573f8f6856cf1bd6cb1e7aa4bf5248b02cbd
Change-Id: Ibc6c9ea45dfbcc2285e3008a2699ad0c096ff2fb