[roll] Roll fuchsia [ffx] Add additional fields to virtual_device.json
The ffx emu tool is building out support for the product bundle
manifest, including a new virtual device specification. This drives
the configuration of the emulator.

This CL adds the fields that previously were defined in the DeviceSpec
struct to the manifest and its schema, and builds on the PBM
functionality to parse the manifest and pass it to the emulator engine.
It also introduces the trait for emulation engines, and embeds the
existing device_launcher-based code into a VDL engine.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604283
Original-Revision: 3ef19247c5ef42255e059e7c94ed5786ed3852d0
GitOrigin-RevId: a213337233e2e78581beacace9e1323a8f0a7cfb
Change-Id: I31dbd50f5869cf1556a54d00ded4b983740db691