[roll] Roll fuchsia [netstack] Add TODOs on empty FIDL structs
Add a TODO on `Empty` FIDL struct definitions referencing the bug
which intends to introduce a built-in empty struct to replace them
when available. In the meantime, the guidance is to continue defining
`Empty` structs wherever we do not expect fields to be added in future;
and to use an anonymous struct otherwise.

While I'm here, fixed a typo.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604726
Original-Revision: b832705d1d0df4bb85b135c8508109fe48a74885
GitOrigin-RevId: 8cc90df96579f917c928ecb14e03b1ef215e0f18
Change-Id: I38a4c510a8a2afa60c5b30da747b25086bfb39b1