[roll] Roll fuchsia [ffx] Add a README to document config_types design
In our redesign of the ffx emu tool, we are introducing a strict
separation of roles in the data types used by the emulation engine
interface. Those roles are expected to be clear to a regular user, but a
developer who's less familiar with the system may be confused by our
intentions.

This CL adds a README document in the config_types crate which explains
the design philosophy around the types and their roles. We expect this
will reduce churn during code reviews in the future, by providing
contributors with a solid foundation to understand the design.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605303
Original-Revision: 09b5618d60ceaa231ad37546ab2afc036b62851e
GitOrigin-RevId: 0ccbafa953b6f8e2db0f0eb6ea3dccb34dd81a4d
Change-Id: Ie54b17951b8ebf8b4b73ef4b4031711071adc3bb