[roll] Roll fuchsia [ffx] Rename the common lib build label
In our recent refactor, we renamed all of the build targets in the ffx
emu plugin so that the build label matches the crate label. We left out
the common lib, because we had another plugin which depended on it. Now
that that dependency is removed, we can finish the refactor safely.

This CL renames the build target fro "lib" to "ffx_emulator_common", to
match the crate label. It also updates all of the references to that
target accordingly.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605304
Original-Revision: 8ddff2d816815bd20c93701c20df5a2ebde05aff
GitOrigin-RevId: d364f1603dc1c01026a44ad07acba5e8310873e4
Change-Id: Id98c9ce24ff2348c18d2d567891cc7017ff68d15