[roll] Roll fuchsia [llcpp] Wire through transport-specific context
The driver transport needs to pass an arena from user-facing
method calls down to transport write calls and in the other
direction from transport read calls to user-facing method
responses.

This CL introduces mechanism to pass through a transport-specific
pointer through these layers.

No test is included in this CL, but it will be covered in an
upcoming integration test. The reason for not adding the integration
test now is that more work needs to be done elsewhere to get
to the point where an integration test is feasible.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/604742
Original-Revision: b25b30f652dd892280cc78a3a0a5a7c6eaa060f5
GitOrigin-RevId: afa7dde2f3f6ba4b44308b7560c80dcbb794567a
Change-Id: If95354f5bbf9806469674679f14897efb62963c7