[roll] Roll fuchsia [gpio][pl061] Delete unused driver
This "driver" doesn't have bind rules, and appears incomplete. The board
it was originally for (hikey960?) was removed from the fuchsia repo a
while ago. I'm avoiding adding this driver to the epitaphs yaml because
I believe it would be more worthwhile to rewrite this driver from
scratch than revive it.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605827
Original-Revision: 798ad5ea534159d74428d67e41acdb27d56facfa
GitOrigin-RevId: fd50821c6899dbe8bc79e9f24823869e98684892
Change-Id: I6427ed1ee4f279f30cd02654d4e3002932dbf138