[roll] Roll fuchsia [cflags] Make FIT_DECLARE_THREAD_CHECKER -Wextra-semi compliant
The previous approach of surrounding the empty form of this macro
doesn't actually work, because the trailing semicolon ends up being
outside the pragmas.
Therefore it's required to include the semicolon in the non-empty
definition of this macro, and to remove the semicolon from all usages.

Upsetting but apparently unavoidable.

Original-Bug: 56257
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605824
Original-Revision: 1760f229dc46fa8e19f75c1b0ea32f50914b61e0
GitOrigin-RevId: ceb4fceb303da24579b48b2e57ddb6955aad18b2
Change-Id: I5c6d3e58a3c64a38516d1b3b9838f67d9f789e59