[roll] Roll fuchsia [virtualization] Update termina build script
This build updates the sonames for EGL and GLES to be
compatible with ldconfig in the container.

Also minigbm is built from source.  We're applying patches
for now but expect to land these after consultation with minigbm
owners.

A magma source area serrves to build libmagma there, making it
available in the sysroot so mesa and minigbm can both link it.

Original-Bug:82886

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605823
Original-Revision: 29e11b666684ae45a4b2c0292dd351e91fae4a97
GitOrigin-RevId: 547617fe321d4cfe5cdd317c94d2ecee14327b3e
Change-Id: I143277918b03d69605303d15606026217968112f