[roll] Roll fuchsia [base-resolver] Disallow package hashes in URL
Currently package hashes are ignored, which can lead to unexpected
package resolution.

This change causes base resolver to return an error when resolving a
component URL with a package hash.

In the future, it might be useful to resolve URLs with package hashes if
they match the package's hash in base, but this isn't needed right now.

Original-Bug: 78072
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605503
Original-Revision: 9d37e028c5b8803c4f6c7059cbfaa17624dc5431
GitOrigin-RevId: 0e6f9a97095246ee73c9d3caccd6d936490489c6
Change-Id: If79f5ce4108c315f0f2904fbd06b729c258cdc88