[roll] Roll fuchsia [archivist][tests] rename "add_eager_component" to "add_eager_child"
The realm builder library is now using "child" instead of "component".
The "add_component" function was renamed to "add_child" in a prior
commit, but "add_eager_component" was missed in this migration.

Rename it to align with the realm builder library it's wrapping.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605902
Original-Revision: 4635daf3a7cc45f20dd082a08ccea4f64eaee02b
GitOrigin-RevId: 9cdf2846f5fe3ac7c3ae71f56a1a4f0cc03e0fae
Change-Id: Icbee349a5a864d92bb839c4df709047986dc010d