[roll] Roll fuchsia [fshost] Add admin privileges to fshost remote connections.
Adding the admin permissions for fshost, as this will allow for shutdown
calls to be made from that connection.

Additionally added a config value to drop the admin flag on calls to
MountData. This is flagged on for f2fs and fxfs.

Original-Bug: 88400
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/605641
Original-Revision: fa87d68868e541ae9c00948c2d514d7ebc6874ff
GitOrigin-RevId: d8f448f351f2817719a3a07d96219c28da72147e
Change-Id: Ica17b3405b3198f801f20077938bbee83e64df0f