[roll] Roll experiences [Ermine][Quicksettings] Add WiFi UI to connect to all security types
Expands WiFi UI to connect to all security types. When an open network is selected, a prompt asking to connect to the selected network is shown. If a password protected network is selected, a prompt to enter a password and connect is shown.

Original-Bug: 79885

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/experiences/+/603451
Original-Revision: 80e6cccc805b3e1844f44f4701c720f99d7c487e
GitOrigin-RevId: e12aa70a65e48b10de4bd909090642f52a755f77
Change-Id: I3f91a23d6c90619f12ee24c243f18ce379a1aa27